Polarizabilities

Alphas is a module to load polarizabilities. In this module, you will find functions to compute electric and magnetic polarizabilities of several objects. The modules also contains some functions to renormalize polarizabilities and to manage the multiple formats of the polarizabilities (dispatch). The list of functions (as well as how to use them) is given below. Note that $N$ denotes the number of point dipoles in the problem.

Format of the polarizabilities in the DDACore and PostProcessing functions

The functions of the DDACore and PostProcessing modules come with an automatic dispatch of the format of the polarizability. If we have $N$ dipoles, it can be:

a complex scalar

a 1D complex array of size $N$

a $3\times 3$ or $6\times 6$ complex matrix.

a 3D complex array of size $N\times 3\times 3$ or $N\times 6\times 6$

Functions Documentation

Main.CoupledElectricMagneticDipoles.Alphas.alpha0_parallelepiped — Method

alpha0_parallelepiped(lx,ly,lz,eps,eps_h)

Computes the quasistatic polarizability tensor of a parallelepiped of dimensions lx,ly,lz and dielectric constant eps in a medium with dielectric constant eps_h. Outputs a $3\times 3$ float matrix with units of volume.

Main.CoupledElectricMagneticDipoles.Alphas.alpha0_sphere — Method

alpha0_sphere(a,eps,eps_h)

Computes the quasistatic polarizability of a sphere of radius a and dielectric constant eps in a medium with dielectric constant eps_h. Outputs a float with units of volume.

Main.CoupledElectricMagneticDipoles.Alphas.alpha0_volume — Method

alpha0_volume(V,eps,eps_h)

Computes the quasistatic polarizability of any object with volume V and dielectric constant eps in a medium with dielectric constant eps_h. Outputs a float with units of volume.

Main.CoupledElectricMagneticDipoles.Alphas.alpha_radiative — Method

alpha_radiative(alpha0,knorm)

Applies the radiative correction to the polarizability matrix or scalar alpha0 (with units of volume). Outputs a (3x3) complex dimensionless scalar or matrix.

Main.CoupledElectricMagneticDipoles.Alphas.alpha_e_m_mie — Method

alpha_e_m_mie(ka,eps,eps_h;mu=1,mu_h=1)

Computes the electric and magnetic polarizabilities from the mie coefficients $a_1$ and $b_1$ of a particle with dimensionless radius ka, and of dielectric permittivity and magnetic permeability eps and mu, in a host medium with dielectric permittivity and magnetic permeability eps_h and mu_h. Outputs two dimensionless scalars that are respectively the electric and the magnetic polarizability.

Main.CoupledElectricMagneticDipoles.Alphas.dispatch_e — Method

dispatch_e(alpha_e_dl,n_particles)

Creates an iterable with the polarizability of all particles in order to facilitate the syntaxis for multuplying a Green's function with the polarizability of a particle i.

Arguments

alpha_e_dl: electric polarizability, see the Alphas module's documentation for the accepted formats.

n_particles: number of particles (integer)

Outputs

alp_e: iterable electric polarizability

Main.CoupledElectricMagneticDipoles.Alphas.dispatch_e_m — Method

dispatch_e_m(alpha_e_dl,alpha_m_dl,n_particles)

Creates an iterable with the polarizability of all particles in order to facilitate the syntaxis for multuplying a Green's function with the polarizability of a particle i.

Arguments

alpha_e_dl: electric polarizability, see the Alphas module's documentation for the accepted formats.

alpha_m_dl: magnetic polarizability, see the Alphas module's documentation for the accepted formats.

n_particles: number of particles (integer)

Outputs

alp_e: iterable electric polarizability

alp_m: iterable magnetic polarizability

Main.CoupledElectricMagneticDipoles.Alphas.dispatch_e_m — Method

dispatch_e_m(alpha_dl,n_particles)

Creates an iterable with the polarizability of all particles in order to facilitate the syntaxis for multuplying a Green's function with the polarizability of a particle i

Arguments

alpha_dl: polarizability 6x6 complex matrix, see the Alphas module's documentation for the accepted formats.

n_particles: number of particles (integer)

Outputs

alp: iterable polarizability

Main.CoupledElectricMagneticDipoles.Alphas.renorm_alpha — Method

renorm_alpha(knorm,alpha)

Renormalizes any polarizability alpha with units of volume in a dimensionless polarizability by multiplying by $k^3/4\pi$. knorm is the wave number in the medium.

Main.CoupledElectricMagneticDipoles.Alphas.denorm_alpha — Method

denorm_alpha(knorm,alpha)

Denormalizes any dimensionless polarizability alpha in a polarizability with units of volume by multiplying by $4\pi /k^3$. knorm is the wave number in the medium.