DDA Core Functionalities

DDACore is the modules allowing to solve CEMD problems (see supplemetary informations in the "Theory" tab for definition of the problem). The problems can be solved either on the CPU (in parrallel with LAPACK) or on the GPU (With CUDA). The list of functions and how to use them is in the next section.

Functions List and Documentation

Main.CoupledElectricMagneticDipoles.DDACore.solve_DDA_e — Method

solve_DDA_e(kr,alpha_e_dl;input_field=nothing,solver="CPU",verbose=true)

Builds and solves the DDA equations under a given input field $\vec{E}_{0}$, i.e.

\[\vec{E}_{i}=\vec{E}_{0}(\vec{r}_{i})+\sum^N_{j\neq i}G_e(\vec{r}_i,\vec{r}_j)\alpha_j\vec{E}_j \]

for a group of $N$ only electric dipoles and returns the incident fields on each of the dipoles.

#Arguments

kr: 2D float array of size $N\times 3$ containing the dimentionless positions $k\vec{r}$ of each dipole.

alpha_e_dl: complex dimensionless electric polarisability of each dipoles. See the Alphas module for accepted formattings.

input_field: 2D complex array of size $N\times 3$ containing the input field $E_0$ at each of the dipoles positions.

solver: string that contains the name of the solver that need to be used. For this, check the DDACore.solve_system function documentation. By default set to "CPU".

verbose: whether to output informations to the standard output during running or not. By default set to true.

#Outputs

e_inc: 2D complex array of size $N\times 3$ containing the incident fields $E_inc$ on every dipole.

Main.CoupledElectricMagneticDipoles.DDACore.solve_DDA_e_m — Method

solve_DDA_e_m(kr,alpha_e_dl,alpha_m_dl;input_field=nothing,solver="CPU",verbose=true)

Builds and solves the DDA equations with dimensionless inputs under a given input field, i.e.

\[\vec{E}_{i} =\vec{E}_{0}\left(\vec{r}_{i}\right)+\sum_{j\neq i}G_{E}\left(\vec{r}_{i},\vec{r}_{j}\right)\alpha_{E}^{(j)}\vec{E}_{j}+iG_{M}\left(\vec{r}_{i},\vec{r}_{j}\right)\alpha_{M}^{(j)}\vec{H}_{j}\]

\[\vec{H}_{i} =\vec{H}_{0}\left(\vec{r}_{i}\right)+\sum_{j\neq i}-iG_{M}\left(\vec{r}_{i},\vec{r}_{j}\right)\alpha_{E}^{(j)}\vec{E}_{j}+G_{E}\left(\vec{r}_{i},\vec{r}_{j}\right)\alpha_{M}^{(j)}\vec{H}_{j}\]

for a group of $N$ electric and magnetic dipoles and return the polarisations of every particle and incident fields on every particle.

#Arguments

kr: 2D float array of size $N\times 3$ containing the dimensionless positions $k\vec{r}$ of all the dipoles.

alpha_e_dl: complex dimensionless electric polarisability of each dipoles. See the Alphas module for accepted formats.

alpha_m_dl: complex dimensionless magnetic polarisability of each dipoles. See the Alphas module for accepted formats.

input_field: 2D complex array of size $N\times 6$ containing the input field $\phi_{0}=(E_0,H_0)$ at each of the dipoles positions.

solver:string that contains the name of the solver that need to be used. For this, check the DDACore.solve_system function documentation. By default set to "CPU".

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

#Outputs

e_inc: 2D complex array of size $N\times 6$ containing the incident fields $\phi_{inc}=(E_inc,H_inc)$ on every dipole.

Main.CoupledElectricMagneticDipoles.DDACore.solve_DDA_e_m — Method

function solve_DDA_e_m(kr,alpha_dl;input_field=nothing,solver="CPU",verbose=true)

Similar to solve_DDA_e_m(kr,alpha_e_dl,alpha_m_dl;input_field=nothing,solver="CPU",verbose=true), but the polarisabilities are $6\times 6$ complex tensors taking into account both electrric and magnetic behaviour of the particles and its optical activity. alpha is a 3D complex array of size $N\times 6\times 6$containing the dimesnionless polarisability of each particle.

Main.CoupledElectricMagneticDipoles.DDACore.solve_system — Method

 solve_system(A,b,solver,verbose)

Solves a system of equations of the type $Ax=b$ where matrix is $A$ and vector is $b$ using the method solver and returns x. x can be a 1D column vector or a 2D matrix. In this second case, the system is going to solve each column of the matrix as a different problem (without re-inverting A). The solver flag can be set to

CPU: In this case, the system is solved using LAPACK on the CPU.

GPU: In this case, the system is solved using CUSOLVE on the GPU (if available).

Main.CoupledElectricMagneticDipoles.DDACore.load_dda_matrix_e — Method

 load_dda_matrix_e(kr,alpha_e_dl,verbose)

Builds the electric only DDA matrix $A=[I-G\alpha]$ with dimensionless postitions kr (2D array of size $Nx3$) and dimensionless polarisabilities alpha_e_dl (see foramt rules in Alphas module). Returns $3N\times 3N$ complex DDA matrix.

Main.CoupledElectricMagneticDipoles.DDACore.load_dda_matrix_e_m — Method

 load_dda_matrix_e_m(kr,alpha_e_dl,alpha_m_dl,verbose)

Builds the electric and magnetic DDA matrix $A=[I-G\alpha]$ with dimensionless postitions kr (2D array of size $N\times 3$) and dimensionless electric and magnetic polarisabilities alpha_e_dl and alpha_m_dl (see format rules in the Alphas module). Return $6N\times 6N$ complex DDA matrix.

Main.CoupledElectricMagneticDipoles.DDACore.load_dda_matrix_e_m — Method

 load_dda_matrix_e_m(kr,alpha_tensor,verbose)

Builds the electric and magnetic DDA matrix $A=[I-G\alpha]$ with dimensionless postitions kr (two dimensional arrays of size $N\times 3$) and dimensionless polarisability alpha_tensor (a three-dimensional complex array of size $N\times 6\times 6$containing the polarisability $6\times 6$ tensor of each dipole). Return $6N\times 6N$ complex DDA matrix