Geometries library

Geometries is a small modules allowing to discretize primitives in small cubes that can be later used in the DDA. The two primitives that are proposed are a sphere and a cube. The description of the two functions are given in the next section. Note that the discretization of the sphere implements anti-aliasing (see Wikipedia page for anti-aliasing or the PS sphere example).

Functions Documentation

Main.CoupledElectricMagneticDipoles.Geometries.discretize_sphere — Function

discretize_sphere(a,N;N_sub=10)

Discretizes the volume of a sphere of radius a in small cubes of edge dx=2*a/N. N_sub is a parameter to set the anti-aliasing accuracy of the discretisation. It is by default set to 10. If $N_d$ is the obtained number of cubes, the output is an $N_d \times 4$ array containing the 3D position of the centers of the cubes and their filling fraction. Returns as well the size of the edge of the cubes dx.

Main.CoupledElectricMagneticDipoles.Geometries.discretize_cube — Method

discretize_cube(L,N)

Discretizes the volume of a cube of edge L in small cubes of edge dx=L/N. If $N_d$ is the obtained number of cubes, the output is an $N_d \times 4$ array containing the 3D position of the centers of the cubes and their filling fraction (i.e. 1). Returns as well the size of the edge of the cubes dx.