Green Tensors

GreenTensors is aimed to compute the renormalized electric and magnetic green tensors. To learn more about the renormalization and the expressions that are used, see the theory part.

Functions Documentation

Renormalized Green Tensors

Main.CoupledElectricMagneticDipoles.GreenTensors.G_e_renorm — Function

G_e_renorm(kr1,kr2)

Computes the renormalized electric green tensor between two position r1 and r2, where the imputs are the positions multiplied by the wave number kr1 and kr2. The output is a dimensionless 3x3 complex matrix.

Main.CoupledElectricMagneticDipoles.GreenTensors.G_m_renorm — Function

G_m_renorm(kr1,kr2)

Computes the renormalized magnetic green tensor between two position r1 and r2, where the imputs are the positions multiplied by the wave number kr1 and kr2. The output is a dimensionless 3x3 complex matrix.

Main.CoupledElectricMagneticDipoles.GreenTensors.G_em_renorm — Function

G_em_renorm(kr1,kr2)

Computes the renormalized electric and magnetic green tensors between two position r1 and r2, where the imputs are the positions multiplied by the wave number kr1 and kr2. The outputs are two dimensionless 3x3 complex matrix.

Main.CoupledElectricMagneticDipoles.GreenTensors.G_em_far_field_renorm — Function

G_em_far_field_renorm(kr1,kr2)

Computes the renormalized electric and magnetic green tensors in the far field approximation between two dimensionless position kr1 and kr2. Note that is is only valid for $kr_1>>kr_2$ and $kr_1>>1$ The outputs are two dimensionless 3x3 complex matrix.

Derivatives of the Renormalized Green Tensors

Main.CoupledElectricMagneticDipoles.GreenTensors.dxG_e_renorm — Function

dxG_e_renorm(kr1,kr2)

Computes the derivative of the renormalized electric green tensor (defined in G_e_renorm(kr1,kr2)) regarding the k*x component of kr1 between two position kr1 and kr2. The output is a dimensionless 3x3 complex matrix.

Main.CoupledElectricMagneticDipoles.GreenTensors.dyG_e_renorm — Function

dyG_e_renorm(kr1,kr2)

Computes the derivative of the renormalized electric green tensor (defined in G_e_renorm(kr1,kr2)) regarding the k*y component of kr1 between two position kr1 and kr2. The output is a dimensionless 3x3 complex matrix.

Main.CoupledElectricMagneticDipoles.GreenTensors.dzG_e_renorm — Function

dzG_e_renorm(kr1,kr2)

Computes the derivative of the renormalized electric green tensor (defined in G_e_renorm(kr1,kr2)) regarding the k*z component of kr1 between two position kr1 and kr2. The output is a dimensionless 3x3 complex matrix.

Main.CoupledElectricMagneticDipoles.GreenTensors.dxG_m_renorm — Function

dxG_m_renorm(kr1,kr2)

Computes the derivative of the renormalized magnetic green tensor (defined in G_m_renorm(kr1,kr2)) regarding the k*x component of kr1 between two position kr1 and kr2. The output is a dimensionless 3x3 complex matrix.

Main.CoupledElectricMagneticDipoles.GreenTensors.dyG_m_renorm — Function

dyG_m_renorm(kr1,kr2)

Computes the derivative of the renormalized magnetic green tensor (defined in G_m_renorm(kr1,kr2)) regarding the k*y component of kr1 between two position kr1 and kr2. The output is a dimensionless 3x3 complex matrix.

Main.CoupledElectricMagneticDipoles.GreenTensors.dzG_m_renorm — Function

dzG_m_renorm(kr1,kr2)

Computes the derivative of the renormalized magnetic green tensor (defined in G_m_renorm(kr1,kr2)) regarding the k*z component of kr1 between two position kr1 and kr2. The output is a dimensionless 3x3 complex matrix.

Main.CoupledElectricMagneticDipoles.GreenTensors.dxG_em_renorm — Function

dxG_em_renorm(kr1,kr2)

Computes the derivative of the renormalized electric and magnetic green tensors (defined in G_e_renorm(kr1,kr2) and G_m_renorm(kr1,kr2)) regarding the k*x component of kr1 between two position kr1 and kr2. The outputs are two dimensionless 3x3 complex matrix.

Main.CoupledElectricMagneticDipoles.GreenTensors.dyG_em_renorm — Function

dyG_em_renorm(kr1,kr2)

Computes the derivative of the renormalized electric and magnetic green tensors (defined in G_e_renorm(kr1,kr2) and G_m_renorm(kr1,kr2)) regarding the k*y component of kr1 between two position kr1 and kr2. The outputs are two dimensionless 3x3 complex matrix.

Main.CoupledElectricMagneticDipoles.GreenTensors.dzG_em_renorm — Function

dzG_em_renorm(kr1,kr2)

Computes the derivative of the renormalized electric and magnetic green tensors (defined in G_e_renorm(kr1,kr2) and G_m_renorm(kr1,kr2)) regarding the k*z component of kr1 between two position kr1 and kr2. The outputs are two dimensionless 3x3 complex matrix.

Utilities Functions

Main.CoupledElectricMagneticDipoles.GreenTensors.denormalize_G_e — Function

denormalize_G_e(Ge,knorm)

Passes from a dimensionless electric green tensor Ge to a green tensor with units of length⁻¹. knorm is the wavenumber into the medium.

Main.CoupledElectricMagneticDipoles.GreenTensors.denormalize_G_m — Function

denormalize_G_m(Gm,knorm)

Passes from a dimensionless magnetic green tensor Gm to a green tensor with units of length⁻². knorm is the wavenumber into the medium.