Input Fields

This module allows to compute input fields. It also implements derivatives of some of the beams. See the theory part to know more about the expression used tu compute the input fields. In general $N$ represents the number of positions inputed in the function.

Functions Documentation

Plane Waves

Main.CoupledElectricMagneticDipoles.InputFields.plane_wave_e — Function

plane_wave_e(krf;khat=[0,0,1],e0=[1,0,0])

Computes a simple plane with dimensionless input evaluated at krf. khat is the direction of propagation and e0 is the polarization. krf is a $N\times 3$ float array. The output is a $N\times 3$ complex array representing the electric field.

Main.CoupledElectricMagneticDipoles.InputFields.plane_wave_e_m — Function

plane_wave_e_m(krf;khat=[0,0,1],e0=[1,0,0])

Computes a simple plane with dimensionless input evaluated at krf. khat is the direction of propagation and e0 is the polarization. krf is a $N\times 3$ float array. The output is a $N\times 6$ complex array representing the electric and magnetic fields.

Point Dipoles Sources

Main.CoupledElectricMagneticDipoles.InputFields.point_dipole_e — Function

point_dipole_e(krf, krd, dip, e0_const=1)

Computes the electric field emitted by a point dipole.

Arguments

krf: 2D float array of size $N\times 3$ containing the dimensionless positions $k\mathbf{r}$ where the field is computed.

krd: 1D float array of size 3 containing the dimensionless position $k\mathbf{r_d}$ where source is located.

dip: integer defining the dipole moment (dip = 1 is an electric x-dipole, dip = 2 an elctric y-dipole...) or complex array of size 6 with the desired dipole moment of the dipole.

e0: scalar with the modulus of the dipole moment.

Outputs

e_dipole:2D $N\times 3$ complex array with the electromagnetic field.

Main.CoupledElectricMagneticDipoles.InputFields.point_dipole_e_m — Function

point_dipole_e_m(krf, krd, dip, e0=1)

Computes the electric and magnetic fields emitted by a point dipole.

Arguments

krf: 2D float array of size $N\times 3$ containing the dimensionless positions $k\mathbf{r}$ where the field is computed.

krd: 1D float array of size 3 containing the dimensionless position $k\mathbf{r_d}$ where source is located.

dip: integer defining the dipole moment (dip = 1 is an electric x-dipole, dip = 2 an elctric y-dipole...) or complex array of size 6 with the desired dipole moment of the dipole.

e0: scalar with the modulus of the dipole moment.

Outputs

phi_dipole: 2D $N\times 6$complex array with the electromagnetic fields at each position.

Gaussian, Hermite-Gaussian and Laguerre-Gaussian Beams

Main.CoupledElectricMagneticDipoles.InputFields.gauss_beam_e — Function

gauss_beam_e(rf,knorm,bw0; e0 = 1, kmax = nothing, maxe = Int(5e3))

Computes the electric field distribution of a Gaussian beam that propagates along the z-axis and where the electric field is polarized along the x-axis (polarized electric). For another polarization just rotate the field in the xy-plane.

Arguments

rf: 2D float array of size $N\times 3$ containing the positions (units of length) where the field is calculated.

knorm: wavenumber in the medium.

bw0: float with the beam waist radius.

e0: scalar with the modulus of the electric field at the origin of coordinates of the theoretical field (including evanescent waves).

kmax: float setting the limit of the radial integration (it shoud be kmax < k).

maxe: maximum number of evaluations in the adapative integral (see Cubature.jl for more details).

Outputs

eh_gauss: 2D complex array of size $N\times 6$ with the value of the field at each position.

Main.CoupledElectricMagneticDipoles.InputFields.ghermite_beam_e — Function

ghermite_beam_e(rf,knorm,bw0,n,m; e0 = 1, kmax = nothing, maxe=Int(5e3))

Computes the electric field distribution of a Hermite-Gaussian beam that propagates along the z-axis and where the electric field is polarized along the x-axis (polarized electric). For another polarization just rotate the field in the xy-plane.

Arguments

rf: 2D float array of size $N\times 3$ containing the positions (units of length) where the field is calculated.

knorm: wavenumber in the medium.

bw0: float with the beam waist radius.

n: int with the order of the beam.

m: int with the degree of the beam.

e0: scalar with the modulus of the electric field (see theory for more details).

kmax: float setting the limit of the radial integration (it shoud be kmax < k).

maxe: maximum number of evaluations in the adapative integral (see Cubature.jl for more details).

Outputs

eh_hermite: 2D complex array of size $N\times 6$ with the value of the field ateach position.

Main.CoupledElectricMagneticDipoles.InputFields.glaguerre_beam_e — Function

glaguerre_beam_e(rf,knorm,bw0,n,m; e0 = 1, kmax = nothing, maxe=Int(5e3))

Computes the electric field distribution of a Laguerre-Gaussian beam that propagates along the z-axis and where the electric field is polarized along the x-axis (polarized electric). For another polarization just rotate the field in the xy-plane.

Arguments

rf: 2D float array of size $N\times 3$ containing the positions (units of length) where the field is calculated.

knorm: wavenumber in the medium.

bw0: float with the beam waist radius.

n: non-negative int with the radial order of the beam.

m: int with the azimuthal order of the beam.

e0: scalar with the modulus of the electric field (see theory for more details).

kmax: float setting the limit of the radial integration (it shoud be kmax < k).

maxe: maximum number of evaluations in the adapative integral (see Cubature.jl for more details).

Outputs

eh_laguerre: 2D complex array of size $N\times 6$ with the value of the field at each position.

Main.CoupledElectricMagneticDipoles.InputFields.gauss_beam_e_m — Function

gauss_beam_e_m(rf, knorm, bw0; e0 = 1, kmax = nothing, maxe = Int(5e3))

Computes the electromagnetic field distribution of a Gaussian beam that propagates along the z-axis and where the electric field is polarized along the x-axis (polarized electric). For another polarization just rotate the field in the xy-plane. Also, for a polarized magnetic field, exchange E with ZH and H with -E.

Arguments

rf: 2D float array of size $N\times 3$ containing the positions (units of length) where the field is calculated.

knorm: wavenumber in the medium.

bw0: float with the beam waist radius.

e0: scalar with the modulus of the electric field at the origin of coordinates of the theoretical field (including evanescent waves).

kmax: float setting the limit of the radial integration (it shoud be kmax < k).

maxe: maximum number of evaluations in the adapative integral (see Cubature.jl for more details).

Outputs

phi_gauss: 2D complex array of size $N\times 6$ with the value of the field at each position.

Main.CoupledElectricMagneticDipoles.InputFields.ghermite_beam_e_m — Function

ghermite_beam_e_m(rf,knorm,bw0,n,m; e0 = 1, kmax = nothing, maxe=Int(5e3))

Computes the electromagnetic field distribution of a Gaussian-Hermite beam that propagates along the z-axis and where the electric field is polarized along the x-axis (polarized electric). For another polarization just rotate the field in the xy-plane. Also, for a polarized magnetic field, exchange E with ZH and H with -E.

Arguments

rf: 2D float array of size $N\times 3$ containing the positions (units of length) where the field is calculated.

knorm: wavenumber in the medium.

bw0: float with the beam waist radius.

n: int with the order of the beam.

m: int with the degree of the beam.

e0: scalar with the modulus of the electric field (see theory for more details).

kmax: float setting the limit of the radial integration (it shoud be kmax < k).

maxe: maximum number of evaluations in the adapative integral (see Cubature.jl for more details).

Outputs

phi_hermite: 2D complex array of size $N\times 6$ with the value of the field at each position.

Main.CoupledElectricMagneticDipoles.InputFields.glaguerre_beam_e_m — Function

glaguerre_beam_e_m(rf,knorm,bw0,n,m; e0 = 1, kmax = nothing, maxe=Int(5e3))

Computes the electromagnetic field distribution of a Laguerre-Gaussian beam that propagates along the z-axis and where the electric field is polarized along the x-axis (polarized electric). For another polarization just rotate the field in the .xy-plane. Also, for a polarized magnetic field, exchange E with HZ and HZ with -E.

Arguments

rf: 2D float array of size $N\times 3$ containing the positions (units of length) where the field is calculated.

knorm: wavenumber in the medium.

bw0: float with the beam waist radius.

n: non-negative int with the radial order of the beam.

m: int with the azimuthal order of the beam.

e0: scalar with the modulus of the electric field (see theory for more details).

kmax: float setting the limit of the radial integration (it shoud be kmax < k).

maxe: maximum number of evaluations in the adapative integral (see Cubature.jl for more details).

Outputs

phi_laguerre: 2D complex array of size $N\times 6$ with the value of the field at each position.

Derivatives of the Beams

Main.CoupledElectricMagneticDipoles.InputFields.d_plane_wave_e — Function

d_plane_wave_e(kr;khat=[0,0,1],e0=[1,0,0])

Computes the derivatives of an electric field generated with plane_wave_e (the arguments are the same). Outputs three 2D arrays of size $N\times 3$ containing the field derivatives with respect k*x, k*y and k*z.

Main.CoupledElectricMagneticDipoles.InputFields.d_plane_wave_e_m — Function

d_plane_wave_e_m(kr;khat=[0,0,1],e0=[1,0,0])

Computes the derivatives of an electromagnetic field generated with plane_wave_e_m (the arguments are the same). Outputs three 2D arrays of size `N\times 6 containing the field derivatives with respect of k*x, k*y and k*z.

Main.CoupledElectricMagneticDipoles.InputFields.d_point_dipole_e — Function

d_point_dipole_e(krf, krd, dip, e0=1)

Computes the derivatives of an electric field generated with point_dipole_e (the arguments are the same). Outputs three 2D arrays of size $N\times 3$ containing the field derivatives with respect k*x, k*y and k*z.

Main.CoupledElectricMagneticDipoles.InputFields.d_point_dipole_e_m — Function

d_point_dipole_e_m(krf, krd, dip, e0=1)

Computes the derivatives of an electromagnetic field generated with point_dipole_e_m (the arguments are the same). Outputs three 2D arrays of size `N\times 6 containing the field derivatives with respect of k*x, k*y and k*z.

Main.CoupledElectricMagneticDipoles.InputFields.d_gauss_beam_e — Function

d_gauss_beam_e(rf,knorm,bw0; e0 = 1, kmax = nothing, maxe = Int(5e3))

Computes the derivatives of an electric field generated with gauss_beam_e (the arguments are the same). Outputs three 2D arrays of size $N\times 3$ containing the field derivatives with respect k*x, k*y and k*z.

Main.CoupledElectricMagneticDipoles.InputFields.d_gauss_beam_e_m — Function

d_gauss_beam_e_m(rf,knorm,bw0; e0 = 1, kmax = nothing, maxe = Int(5e3))

Computes the derivatives of an electromagnetic field generated with gauss_beam_e_m (the arguments are the same). Outputs three 2D arrays of size `N\times 6 containing the field derivatives with respect of k*x, k*y and k*z.