Mie Coefficients

MieCoeff is a small modules to compute Mie Coefficients, mostly to use it to compute polarizabilities. It also contains some useful function to compute cross sections from the Mie theory.

Expressions are taken in Craig F. Bohren, Donald R. Huffman, Absorption and Scattering of Light by Small Particles (1998).

Functions Documentation

Main.CoupledElectricMagneticDipoles.MieCoeff.mie_an — Method

mie_an(knorm, a, eps, eps_h, n)

Computes the n-th mie coefficient $a_n$ of a sphere of radius a with dielectric constant eps, in a medium with dielectric constant eps_h. knorm is wavenumber in the medium. Retruns a complex scalar.

Main.CoupledElectricMagneticDipoles.MieCoeff.mie_bn — Method

mie_bn(knorm, a, eps, eps_h, n)

Computes the n-th mie coefficient $b_n$ of a sphere of radius a with dielectric constant eps, in a medium with dielectric constant eps_h. knorm is wavenumber in the medium. Returns a complex scalar.

Main.CoupledElectricMagneticDipoles.MieCoeff.mie_scattering_cross_section — Method

mie_scattering_cross_section(knorm,a,eps,eps_h;cutoff=20)

Computes the scattering cross section $C_{sca}$ of a sphere of radius a with dielectric constant eps in a medium with dielectric constant eps_h. knorm is the wavenumber in the medium. For this, we use:

\[C_{sca} =\frac{2\pi}{k^2}\sum^{\infty}_{n=1}\left(2n+1\right)\left(|a_n|^2+|b_n|^2\right)\]

The infinite sum is computed only for terms under the cutoff variable (by default set to 20).

Returns a float with units of surface.

Main.CoupledElectricMagneticDipoles.MieCoeff.mie_extinction_cross_section — Method

mie_extinction_cross_section(knorm,a,eps,eps_h;cutoff=20)

Computes the extinction cross section $C_{sca}$ of a sphere of radius a with dielectric constant eps in a medium with dielectric constant eps_h. knorm is the wavenumber in the medium. For this, we use:

\[C_{ext} =\frac{2\pi}{k^2}\sum^{\infty}_{n=1}\left(2n+1\right)Re\left(a_n+b_n\right)\]

The infinite sum is computed only for terms under the cutoff variable (by default set to 20).

Returns a float with units of surface.

Main.CoupledElectricMagneticDipoles.MieCoeff.mie_absorption_cross_section — Method

mie_absorption_cross_section(knorm,a,eps,eps_h;cutoff=20)

Computes the extinction cross section $C_{sca}$ of a sphere of radius a with dielectric constant eps in a medium with dielectric constant eps_h. knorm is the wavenumber in the medium. For this, we use:

\[C_{abs} =C_{ext}-C_{sca}\]

The infinite sum is computed only for terms under the cutoff variable (by default set to 20).

Returns a float with units of surface.