Post Processing

The PostProcessing modules allows you to process the results of a CEMD calculation. Mostly for:

Computing cross sections.

Computing scattered fields.

Computing local density of states (LDOS).

The list of functions and a description of how to use them is given in the following section.

Functions List and Documentation

Main.CoupledElectricMagneticDipoles.PostProcessing.compute_dipole_moment — Method

compute_dipole_moment(alpha,phi_inc)

Computes the dipole moment (magnetic or electric) of a dipole with polarizability alpha under an incident field phi_inc. alpha can be:

a complex scalar

a 1D complex array of size $N$

a $3\times 3$ or $6\times 6$ complex matrix.

a 3D complex array of size $N\times 3\times 3$ or $N\times 6\times 6$

Main.CoupledElectricMagneticDipoles.PostProcessing.compute_cross_sections_e — Method

compute_cross_sections_e(knorm,kr,e_inc,alpha,input_field;explicit_scattering=true,verbose=true)

Computes the extinction, absorbtion ans scattering cross section $\sigma_{ext}$, $\sigma_{abs}$, $\sigma_{sca}$ of a system made out of electric dipoles. Notes that it sould follow the optical theorem, i.e.

\[\sigma_{ext}=\sigma_{abs}+\sigma_{sca}\]

Inputs

knorm: wavenumber

kr: 2D float array of size $N\times 3$ containing the dimentionless positions $k\vec{r}$ of each dipole.

e_inc: 2D complex array of size $N\times 3$ containing the incident fields $E_{inc}$ on every dipole.

alpha_dl: complex dimensionless electric polarisability of each dipoles. See the Alphas module for accepted formats.

input_field: 2D complex array of size $N\times 3$ containing the input field $E_0$ at each of the dipoles positions.

explicit_scattering: boolean that says wether to compute scttering cross section explicitely (formula) or to deduce it from the optical theorem. By default set to true.

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

Outputs

an array containing in order: extinction, absorption and scattering cross section.

Main.CoupledElectricMagneticDipoles.PostProcessing.compute_cross_sections_e_m — Method

compute_cross_sections_e_m(knorm,kr,phi_inc,alpha_e_dl,alpha_m_dl,input_field;explicit_scattering=true,verbose=true)

Computes the extinction, absorbtion ans scattering cross section $\sigma_{ext}$, $\sigma_{abs}$, $\sigma_{sca}$ of a system made out of electric dipoles. Notes that it sould follow the optical theorem, i.e.

\[\sigma_{ext}=\sigma_{abs}+\sigma_{sca}\]

Inputs

knorm: wavenumber

kr: 2D float array of size $N\times 3$ containing the dimentionless positions $k\vec{r}$ of each dipole.

e_inc: 2D complex array of size $N\times 6$ containing the incident fields $E_{inc}$ on every dipole.

alpha_e_dl: complex dimensionless electric polarisability of each dipoles. See the Alphas module for accepted formats.

alpha_m_dl: complex dimensionless magnetic polarisability of each dipoles. See the Alphas module for accepted formats.

input_field: input field on each dipole ($N\times 6$) array.

explicit_scattering: boolean that says wether to compute scttering cross section explicitely (formula) or to deduce it from the optical theorem. By default set to true.

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

Outputs

an array containing in order: extinction, absorption and scattering cross section.

Main.CoupledElectricMagneticDipoles.PostProcessing.field_sca_e — Method

function field_sca_e(kr, alpha_e_dl, e_inc, krf)

It computes the scattered field from the ensemble of dipoles.

#Arguments

kr: 2D float array of size $N\times 3$ containing the dimentionless positions $k\vec{r}$ of each dipole.

alpha_dl: complex dimensionless electric polarisability of each dipoles. See the Alphas module for accepted formats.

e_inc: 2D complex array of size $N\times 3$ with the incident field $E_{inc}$ in the dipoles.

krf: 2D float array of size $Nf\times 3$ containing the dimentionless positions $k\vec{r_f}$ where the scattered field is calculated.

#Outputs

field_r: 2D complex array of size $Nf\times 3$ with the field scattered by the dipoles at every $k\vec{r_f}$.

Equation

\[\mathbf{E}_{sca}(\mathbf{r}) = k^2G(\mathbf{r},\mathbf{\bar{r}}_N) \alpha(\mathbf{\bar{r}}_N) \mathbf{E}_{inc}(\mathbf{\bar{r}}_N) = k^2 G(\mathbf{r},\mathbf{\bar{r}}_N) \alpha(\mathbf{\bar{r}}_N) D(\mathbf{\bar{r}}_N) \E_{0}\]

\r = r0 \bar{r}}N = pos \alphagg(\mathbf{\bar{r}}N) = alpha \GG(\r,\mathbf{\bar{r}}N) = `Gtensor\E_{inc}(\mathbf{\bar{r}}_N) =E_inc`

\E{sca}(\r) = `fieldr`

Main.CoupledElectricMagneticDipoles.PostProcessing.field_sca_e_m — Method

field_sca_e_m(kr, alpha_e_dl, alpha_m_dl, e_inc, krf)

It computes the scattered field from the ensemble of dipoles.

#Arguments

kr: 2D float array of size $N\times 3$ containing the dimentionless positions $k\vec{r}$ of each dipole.

alpha_e_dl: complex dimensionless electric polarisability of each dipoles. See the Alphas module for accepted formats.

alpha_m_dl: complex dimensionless magnetic polarisability of each dipoles. See the Alphas module for accepted formats.

e_inc: 2D complex array of size $N\times 6$ with the incident field $\phi_{inc}=(E_inc,H_inc)$ on the dipoles.

krf: 2D float array of size $Nf\times 3$ containing the dimentionless positions $k\vec{r_f}$ where the scattered field is calculated.

#Outputs

field_r: 2D complex array of size $Nf\times 6$ with the field scattered by the dipoles at every $k\vec{r_f}$.

Equation

\[\mathbf{E}_{sca}(\mathbf{r}) = k^2G(\mathbf{r},\mathbf{\bar{r}}_N) \alpha(\mathbf{\bar{r}}_N) \mathbf{E}_{inc}(\mathbf{\bar{r}}_N) = k^2 G(\mathbf{r},\mathbf{\bar{r}}_N) \alpha(\mathbf{\bar{r}}_N) D(\mathbf{\bar{r}}_N) \E_{0}\]

\r = r0 \bar{r}}N = pos \alphagg(\mathbf{\bar{r}}N) = alpha \GG(\r,\mathbf{\bar{r}}N) = `Gtensor\E_{inc}(\mathbf{\bar{r}}_N) =E_inc`

\E{sca}(\r) = `fieldr`

Main.CoupledElectricMagneticDipoles.PostProcessing.field_sca_e_m — Method

field_sca_e_m(kr, alpha_dl, e_inc, krf)

It computes the scattered field from the ensemble of dipoles.

#Arguments

kr: 2D float array of size $N\times 3$ containing the dimentionless positions $k\vec{r}$ of each dipole.

alpha_dl: complex dimensionless electric and magnetic polarisability of each dipoles. See the Alphas module for accepted formats.

e_inc: 2D complex array of size $N\times 6$ with the incident field $E_{inc}$ on the dipoles.

krf: 2D float array of size $Nf\times 3$ containing the dimentionless positions $k\vec{r_f}$ where the scattered field is calculated.

#Outputs

field_r: 2D complex array of size $Nf\times 6$ with the field scattered by the dipoles at every $k\vec{r_f}$.

Main.CoupledElectricMagneticDipoles.PostProcessing.ldos_e — Method

ldos_e(kr, alpha_e_dl, Ainv, krd; dip=nothing)

It Computes local density of states (LDOS) by the imaginary part of the returning field.

#Arguments

kr: 2D float array of size $N\times 3$ containing the dimentionless positions $k\vec{r}$ of each dipole.

alpha_e_dl: complex dimensionless electric polarisability of each dipoles. See the Alphas module for accepted formats.

Ainv: (inverse) DDA matrix $[I - G*alpha]^(-1)$.

krd: 2D float array of size $Nd\times 3$ containing the dimentionless positions $k\vec{r_d}$ where the LDOS is calculated.

dip: integer defining the dipole moment ($dip = 1$ is an electric x-dipole, $dip = 2$ an elctric y-dipole...) or float array of size 3 with the desired dipole moment of the dipole.

#Outputs

LDOS: float array with the LDOS.

Equation

\[\mathrm{LDOS}(\mathbf{\bar{r}}_N,\r_0) = 1 + \dfrac{1}{|\bm \mu|^2 }\dfrac{6\pi}{k^3} \Im\left[\bm \mu^{*} \cdot k^2\GG(\r_0,\mathbf{\bar{r}}_N) \alphagg(\mathbf{\bar{r}}_N) \DD(\mathbf{\bar{r}}_N) k^2 \GG(\mathbf{\bar{r}}_N, \r_0) \bm \mu \right]\]

\alphagg(\mathbf{\bar{r}}N) = alpha \DD(\mathbf{\bar{r}}N) = Ainv = [I - k^2Galpha]^(-1) \bar{r}}N = pos \r0 = rd \dfrac{1}{\epsilon0\epsilon} \bm \mu = `dipo` (the pre-factor -\dfrac{1}{\epsilon_0\epsilon}- dessapears after normalization)

k^2\GG(\r0,\mathbf{\bar{r}}N) \alphagg(\mathbf{\bar{r}}N) \DD(\mathbf{\bar{r}}N) k^2 \GG(\mathbf{\bar{r}}N, \r0) \bm \mu = field_r \DD(\mathbf{\bar{r}}N) k^2 \GG(\mathbf{\bar{r}}N, \r0) \bm \mu = `Einc`

\mathrm{LDOS}(\mathbf{\bar{r}}N,\r0) = LDOS

Main.CoupledElectricMagneticDipoles.PostProcessing.ldos_e_m — Method

ldos_e_m(kr, alpha_e_dl, alpha_m_dl, Ainv, krd; dip=nothing)

It Computes local density of states (LDOS) by the imaginary part of the returning field.

#Arguments

kr: 2D float array of size $N\times 3$ containing the dimentionless positions $k\vec{r}$ of each dipole.

alpha_e_dl: complex dimensionless electric polarisability of each dipoles. See the Alphas module for accepted formats.

alpha_m_dl: complex dimensionless magnetic polarisability of each dipoles. See the Alphas module for accepted formats.

Ainv: (inverse) DDA matrix $[I - G*alpha]^(-1)$.

krd: 2D float array of size $Nd\times 3$ containing the dimentionless positions $k\vec{r_d}$ where the LDOS is calculated.

dip: integer defining the dipole moment ($dip = 1$ is an electric x-dipole, $dip = 2$ an elctric y-dipole...) or float array of size 6 with the desired dipole moment of the dipole.

#Outputs

LDOS: float array with the LDOS.

Equation

\[\mathrm{LDOS}(\mathbf{\bar{r}}_N,\r_0) = 1 + \dfrac{1}{|\bm \mu|^2 }\dfrac{6\pi}{k^3} \Im\left[\bm \mu^{*} \cdot k^2\GG(\r_0,\mathbf{\bar{r}}_N) \alphagg(\mathbf{\bar{r}}_N) \DD(\mathbf{\bar{r}}_N) k^2 \GG(\mathbf{\bar{r}}_N, \r_0) \bm \mu \right]\]

\alphagg(\mathbf{\bar{r}}N) = alpha \DD(\mathbf{\bar{r}}N) = Ainv = [I - k^2Galpha]^(-1) \bar{r}}N = pos \r0 = rd \dfrac{1}{\epsilon0\epsilon} \bm \mu = `dipo` (the pre-factor -\dfrac{1}{\epsilon_0\epsilon}- dessapears after normalization)

k^2\GG(\r0,\mathbf{\bar{r}}N) \alphagg(\mathbf{\bar{r}}N) \DD(\mathbf{\bar{r}}N) k^2 \GG(\mathbf{\bar{r}}N, \r0) \bm \mu = field_r \DD(\mathbf{\bar{r}}N) k^2 \GG(\mathbf{\bar{r}}N, \r0) \bm \mu = `Einc`

\mathrm{LDOS}(\mathbf{\bar{r}}N,\r0) = LDOS

Main.CoupledElectricMagneticDipoles.PostProcessing.ldos_e_m — Method

ldos_e_m(kr, alpha_dl, Ainv, krd; dip=nothing)

It Computes local density of states (LDOS) by the imaginary part of the returning field.

#Arguments

kr: 2D float array of size $N\times 3$ containing the dimentionless positions $k\vec{r}$ of each dipole.

alpha_dl: complex dimensionless electric and magnetic polarisability of each dipoles. See the Alphas module for accepted formats.

Ainv: (inverse) DDA matrix $[I - G*alpha]^(-1)$.

krd: 2D float array of size $Nd\times 3$ containing the dimentionless positions $k\vec{r_d}$ where the LDOS is calculated.

dip: integer defining the dipole moment ($dip = 1$ is an electric x-dipole, $dip = 2$ an elctric y-dipole...) or float array of size 6 with the desired dipole moment of the dipole.

#Outputs

LDOS: float array with the LDOS.

Equation

\[\mathrm{LDOS}(\mathbf{\bar{r}}_N,\r_0) = 1 + \dfrac{1}{|\bm \mu|^2 }\dfrac{6\pi}{k^3} \Im\left[\bm \mu^{*} \cdot k^2\GG(\r_0,\mathbf{\bar{r}}_N) \alphagg(\mathbf{\bar{r}}_N) \DD(\mathbf{\bar{r}}_N) k^2 \GG(\mathbf{\bar{r}}_N, \r_0) \bm \mu \right]\]

\alphagg(\mathbf{\bar{r}}N) = alpha \DD(\mathbf{\bar{r}}N) = Ainv = [I - k^2Galpha]^(-1) \bar{r}}N = pos \r0 = rd \dfrac{1}{\epsilon0\epsilon} \bm \mu = `dipo` (the pre-factor -\dfrac{1}{\epsilon_0\epsilon}- dessapears after normalization)

k^2\GG(\r0,\mathbf{\bar{r}}N) \alphagg(\mathbf{\bar{r}}N) \DD(\mathbf{\bar{r}}N) k^2 \GG(\mathbf{\bar{r}}N, \r0) \bm \mu = field_r \DD(\mathbf{\bar{r}}N) k^2 \GG(\mathbf{\bar{r}}N, \r0) \bm \mu = `Einc`

\mathrm{LDOS}(\mathbf{\bar{r}}N,\r0) = LDOS