Polarisabilities

Alphas is a module to load polarizabilities. In this module, you will find functions to compute electric and magnetic polarizabilities of several objects. The modules also contains some functions to renormalize polarizabilities and to manage the multiple formats of the polarizabilities (dispatch). The list of functions as well as how to use them is given below.

Format of the polarizabilities in the DDACore and PostProcessing functions

The functions of the DDACore and PostProcessing modules come with an automatic dispatch of the format of the polarizability. If we have $N$ dipoles, it can be:

a complex scalar

a 1D complex array of size $N$

a $3\times 3$ or $6\times 6$ complex matrix.

a 3D complex array of size $N\times 3\times 3$ or $N\times 6\times 6$

Functions Documentation

Main.CoupledElectricMagneticDipoles.Alphas.alpha0_parallelepiped — Method

alpha0_parallelepiped(lx,ly,lz,eps,eps_h)

Computes the quasistatic polarizability tensor of a parallelepiped of dimensions lx,ly,lz and dielectric constant eps in a medium with dielectric constant eps_h. Outputs a $3\times 3$ float matrix with units of volume.

Main.CoupledElectricMagneticDipoles.Alphas.alpha0_sphere — Method

alpha0_sphere(a,eps,eps_h)

Computes the quasistatic polarizability of a sphere of radiusa and dielectric constant eps in a medium with dielectric constant eps_h. Outputs a float matrix with units of volume.

Main.CoupledElectricMagneticDipoles.Alphas.alpha0_volume — Method

alpha0_volume(a,eps,eps_h)

Computes the quasistatic polarizability of any object with volume V and dielectric constant eps in a medium with dielectric constant eps_h. Outputs a float matrix with units of volume.

Main.CoupledElectricMagneticDipoles.Alphas.alpha_radiative — Method

alpha_radiative(alpha0,knorm)

Applies the radiative correction to the polarizability tensor or scalar alpha0, with units of voume. Outputs a (3x3) complex dimensionless scalar or tensor computed as follow:

\[\alpha=\frac{k^3}{4\pi}\left(\alpha_{0}^{-1}-i\frac{k{{}^3}}{6\pi}\right)^{-1}\]

Main.CoupledElectricMagneticDipoles.Alphas.alpha_e_m_mie — Method

alpha_e_m_mie(vac_knorm,a,eps,eps_h)

Computes the electric and magnetic polarisabilitie from the mie coefficients $a_1$ and $b_1$ of a particle of dielectric constant eps and radius a in a medium with dielectric constant eps_h. knorm is the wavenumber in the medium It outputs two dimensionless scalars which are computed with:

\[\alpha_{E} =\frac{k^{3}\tilde{\alpha}_{E}}{4\pi}=i\frac{3}{2}a_{1},\ \alpha_{M} =\frac{k^{3}\tilde{\alpha}_{M}}{4\pi}=i\frac{3}{2}b_{1}\]

Main.CoupledElectricMagneticDipoles.Alphas.dispatch_e — Method

dispatch_e(alpha_e_dl,n_particles)

Creates an iterable with the polarizability of all particles in order to facilitate the syntaxis for multuply a Green function for the polarizability of particle i

Imputs

alpha_e_dl = electric polarizability

Outputs

alp_e = iterable electric polarizability

Main.CoupledElectricMagneticDipoles.Alphas.dispatch_e_m — Method

dispatch_e_m(alpha_e_dl,alpha_m_dl,n_particles)

Creates an iterable with the polarizability of all particles in order to facilitate the syntaxis for multuply a Green function for the polarizability of particle i.

#Arguments

alpha_e_dl: = electric polarizability

alpha_m_dl: = magnetic polarizability

#Outputs

alp_e: = iterable electric polarizability

alp_m: = iterable magnetic polarizability

Main.CoupledElectricMagneticDipoles.Alphas.dispatch_e_m — Method

dispatch_e_m(alpha_dl,n_particles)

Creates an iterable with the polarizability of all particles in order to facilitate the syntaxis for multuply a Green function for the polarizability of particle i

#Arguments

alpha_dl: = polarizability

#Outputs

alp: = iterable polarizability

Main.CoupledElectricMagneticDipoles.Alphas.renorm_alpha — Method

renorm_alpha(knorm,alpha)

Renormalizes any polarizability with units of volume in a dimensionless polarizability by multiplying by $k^3/4\pi$.

Main.CoupledElectricMagneticDipoles.Alphas.denorm_alpha — Method

denorm_alpha(knorm,alpha)

Denormalizes any dimensionless polarizability in a polarizability with units of volume by multiplying by $4\pi /k^3$.