Polarisabilities

Functions Documentation

Main.CoupledElectricMagneticDipoles.Alphas.depolarisation_tensor — Function

depolarisation_tensor(lx,ly,lz,Vn)

Compute the (3x3) depolarisation tensor $L$ of a rectangular box of sides lx, ly and lz and volume Vn.

Main.CoupledElectricMagneticDipoles.Alphas.alpha_0 — Function

alpha_0(e,e_m,Ln,Vn)

Compute the electric quasistatic polarisabilitiy $\alpha_0$ of a particle with dielectric constant e, volume Vn and depolarisation tensor Ln in a medium with dielectric constant e_m. The output is a (3x3) tensor computed as follow:

\[\alpha_{0}=(\epsilon-\epsilon_{m}I)((\epsilon-\epsilon_{m}I)+L^{-1}\epsilon_{m})^{-1}L^{-1}V\]

Main.CoupledElectricMagneticDipoles.Alphas.alpha_radiative — Function

 alpha_radiative(a0,knorm)

Apply the radiative correction to the polarisability tensor a0 The output is a (3x3) tensor computed as follow:

\[\alpha=\left(\alpha_{0}^{-1}-i\frac{k{{}^3}}{6\pi}\right)^{-1}\]

Main.CoupledElectricMagneticDipoles.Alphas.alpha_e_m_mie — Function

 alpha_e_m_mie(knorm,vac_knorm,a,n,n_m)

Computes the electric and magnetic polarisabilities from the mie coefficients of a particle of refractive indes n and radius a in a medium with refractive index n_m with wavenumber knorm amd vacuum wavenumber vac_knorm. It outputs two scalars which are computed as

\[\tilde{\alpha}_{E} =i\left(\frac{k^{3}}{6\pi}\right)^{-1}a_{1},\ \tilde{\alpha}_{M} =i\left(\frac{k^{3}}{6\pi}\right)^{-1}b_{1}\]

Main.CoupledElectricMagneticDipoles.Alphas.alpha_e_m_mie_renorm — Function

 alpha_e_m_mie_renorm(knorm,vac_knorm,a,n,n_m)

Computes the electric and magnetic polarisabilities in the renormalized system of units from the mie coefficients of a particle of refractive indes n and radius a in a medium with refractive index n_m with wavenumber knorm amd vacuum wavenumber vac_knorm. It outputs two scalars which are computed as

\[\alpha_{E} =\frac{k^{3}\tilde{\alpha}_{E}}{4\pi}=i\frac{3}{2}a_{1},\ \alpha_{M} =\frac{k^{3}\tilde{\alpha}_{M}}{4\pi}=i\frac{3}{2}b_{1}\]