DDA Core Functionalities

Main.CoupledElectricMagneticDipoles.DDACore.solve_DDA_e — Method

 solve_DDA_e(knorm,r,alpha,input_field::Function;solver="LAPACK",verbose=true)

Builds and solves the DDA equations under a given input field, i.e.

\[\vec{E}_{i}=\vec{E}_{0}(\vec{r}_{i})+k^2\sum_{i\neq j}^{N}\tilde{G}_e(\vec{r}_{i},\vec{r}_{j})\alpha_{j}\vec{E}_{j},\]

for a group of $N$ only electric dipoles and returns the polarizations and incident fields of every dipole.

#Arguments

knorm: the wavenumber of the input field.

r: a two-dimensional float array of size $N\times 3$ containing the positions $\vec{r}$ of all the dipoles.

alpha: a three-dimensional complex array of size $N\times 3\times 3$containing the polarisability $\times 3$ tensor of each dipole, or one dimensional array of size $N$ containing the scalar polarizability of each dipole.

input_field: a function taking the wavenumber and one position vector of length 3, i.e. of the form field(knorm,r), and that output a complex array of length 3 which computes the input field $E_0$ evaluated at this position. Can be for example a plane wave or a point source.

solver: a string that contains the name of the solver that need to be used. For this, check the correponding section on the home page. By default set to "LAPACK".

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

#Outputs

p: a two-dimensional complex array of size $N\times 3$ containing the polarization $p$ of each dipole.

e_inc: a two-dimensional complex array of size $N\times 3$ containing the incident fields $E_inc$ on each dipole.

Main.CoupledElectricMagneticDipoles.DDACore.solve_DDA_e — Method

 solve_DDA_e(knorm,r,alpha;solver="JULIA",verbose=true)

Similar to solve_DDA_e(knorm,r,alpha,input_field::Function;solver="LAPACK",verbose=true), but without input field. Returns then the inverse of the DDA Matrix. #Arguments

knorm: the wavenumber of the input field.

r: a two-dimensional float array of size $N\times 3$ containing the positions $\vec{r}$ of all the dipoles.

alpha: a three-dimensional complex array of size $N\times 3\times 3$containing the polarisability $3x3$ tensor of each dipole, or one dimensional array of size $N$ containing the scalar polarizability of each dipole.

solver: string that contains the name of the invertion method that need to be used. For this, check the correponding section on the home page. By default set to "JULIA".

verbose: whether to output informations to the standard output during running or not. By default set to true.

#Output

$3N\times 3N$ inverse of the DDA matrix.

Main.CoupledElectricMagneticDipoles.DDACore.solve_DDA_e — Method

 solve_DDA_e(kr,alpha_dl,input_field::Function;solver="LAPACK",verbose=true)

Builds and solves the DDA equations with dimensionless input under a given input field, i.e.

\[\vec{E}_{i}=\vec{E}_{0}(\vec{r}_{i})+,\sum^N_{j\neqj}G_e(\vec{r}_i,\vec{r}_j)\alpha_j\vec{E}_j \end{equation}\]

for a group of $N$ only electric dipoles and returns the polarizations and incident fields.

#Arguments

kr: two dimensional float array of size $N\times 3$ containing the dimentionless positions $k\vec{r}$ of each dipole.

alpha_dl: three dimensional complex array of size $N\times 3\times 3$containing the dimensionless polarisability $3\times 3$ tensor of each dipole, or one dimenstional array of size $N$ containing the scalar polarizability of each dipole.

input_field: function taking one dimensionless position vector of length 3, i.e. of the form field(kr), and that output a complex array of length 3 which compute the input field $E_0$ elvaluated at this position. Can be for example a plane wave or a point source.

solver: string that contains the name of the solver that need to be used. For this, check the correponding section on the home page. By default set to "LAPACK".

verbose: whether to output informations to the standard output during running or not. By default set to true.

#Outputs

p: two dimensional complex array of size $N\times 3$ containing the polarizations $p$ of each dipoles (in units of the electric field!).

e_inc: two dimensional complex array of size $N\times 3$ containing the incident fields $E_inc$ on every dipole.

Main.CoupledElectricMagneticDipoles.DDACore.solve_DDA_e — Method

 solve_DDA_e(kr,alpha_dl;solver="JULIA",verbose=true)

Similar to solve_DDA_e(kr,alpha_dl,input_field::Function;solver="LAPACK",verbose=true), but without input field. Return then the inverse of the DDA Matrix.

#Arguments

kr: two dimensional float array of size $N\times 3$ containing the dimentionless positions $k\vec{r}$ of each dipole.

alpha_dl: three dimensional complex array of size $N\times 3\times 3$containing the dimensionless polarisability $3\times 3$ tensor of each dipole, or one dimenstional array of size $N$ containing the scalar polarizability of each dipole.

solver: string that contains the name of the solver that need to be used. For this, check the correponding section on the home page. By default set to "LAPACK".

verbose: whether to output informations to the standard output during running or not. By default set to true.

#Output

$3N\times 3N$ inverse of the DDA matrix.

Main.CoupledElectricMagneticDipoles.DDACore.solve_DDA_e_m — Method

 solve_DDA_e_m(knorm,r,alpha_e,alpha_m,input_field::Function;solver="LAPACK",verbose=true)

Builds and solves the DDA equations under a given input field, i.e.

\[\begin{align} \vec{E}_{i} & =\vec{E}_{0}\left(\vec{r}_{i}\right)+\sum_{j\ne i}\mathbb{G}_{E}\left(\vec{r}_{i},\vec{r}_{j}\right)\alpha_{E}^{(j)}\vec{E}_{j}+i\mathbb{G}_{M}\left(\vec{r}_{i},\vec{r}_{j}\right)\alpha_{M}^{(j)}\vec{H}_{j}\label{eq:100a}\\ \vec{H}_{i} & =\vec{H}_{0}\left(\vec{r}_{i}\right)+\sum_{j\ne i}-i\mathbb{G}_{M}\left(\vec{r}_{i},\vec{r}_{j}\right)\alpha_{E}^{(j)}\vec{E}_{j}+\mathbb{G}_{E}\left(\vec{r}_{i},\vec{r}_{j}\right)\alpha_{M}^{(j)}\vec{H}_{j}\label{eq:100b} \end{align}\]

for a group of $N$ electric and magnetic dipoles and return the polarisations of every particle and incident fields on every particle.

#Arguments

knorm: the wavenumber of the input field.

r: a two-dimensional float array of size $N\times 3$ containing the positions $\vec{r}$ of all the dipoles.

alpha_e: a three-dimensional complex array of size $N\times 3\times 3$containing the electric polarisability $3\times 3$ tensor of every dipole, or one dimensional array of size $N$ containing the scalar polarizability of every dipole.

alpha_m: a three-dimensional complex array of size $N\times 3\times 3$containing the magnetic polarisability $3\times 3$ tensor of every dipole, or one dimensional array of size $N$ containing the scalar polarizability of every dipole.

input_field: a function taking the wavenumber and one position vector of length 3, i.e. of the form field(knorm,r), and that output a complex array of length 3 which computes the input fields $E_0$ and $H_0$ evaluated at this position. Can be for example a plane wave or a point source.

solver: a string that five the name of the solver that need to be used. For this, check the correponding section on the home page. By default set to "LAPACK".

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

#Outputs

p: a two-dimensional complex array of size $N\times 3$ containing the electric polarizations $p$ moments of all the electric dipoles.

m: a two-dimensional complex array of size $N\times 3$ containing the magnetic polarizations $m$ moments of all the magnetic dipoles.

e_inc: a two-dimensional complex array of size $N\times 3$ containing the incident electric fields $E_inc$ on all the particles.

h_inc: a two-dimensional complex array of size $N\times 3$ containing the incident magnetic fields $H_inc$ on all the particles.

e_inp: a two-dimensional complex array of size $N\times 3$ containing the input electric fields $E_inc$ on all the particles.

h_inp: a two-dimensional complex array of size $N\times 3$ containing the input magnetic fields $H_inc$ on all the particles.

Main.CoupledElectricMagneticDipoles.DDACore.solve_DDA_e_m — Method

 solve_DDA_e_m(knorm,r,alpha_e,alpha_m;solver="LAPACK",verbose=true)

Builds and solves the DDA equations under a given input field, i.e.

\[\begin{align} \vec{E}_{i} & =\vec{E}_{0}\left(\vec{r}_{i}\right)+\sum_{j\ne i}\mathbb{G}_{E}\left(\vec{r}_{i},\vec{r}_{j}\right)\alpha_{E}^{(j)}\vec{E}_{j}+i\mathbb{G}_{M}\left(\vec{r}_{i},\vec{r}_{j}\right)\alpha_{M}^{(j)}\vec{H}_{j}\label{eq:100a}\\ \vec{H}_{i} & =\vec{H}_{0}\left(\vec{r}_{i}\right)+\sum_{j\ne i}-i\mathbb{G}_{M}\left(\vec{r}_{i},\vec{r}_{j}\right)\alpha_{E}^{(j)}\vec{E}_{j}+\mathbb{G}_{E}\left(\vec{r}_{i},\vec{r}_{j}\right)\alpha_{M}^{(j)}\vec{H}_{j}\label{eq:100b} \end{align}\]

for a group of $N$ electric and magnetic dipoles and return the polarisations of every particle and incident fields on every particle.

#Arguments

knorm: the wavenumber of the input field.

r: a two-dimensional float array of size $N\times 3$ containing the positions $\vec{r}$ of all the dipoles.

alpha_e: a three-dimensional complex array of size $N\times 3\times 3$containing the electric polarisability $3\times 3$ tensor of every dipole, or one dimensional array of size $N$ containing the scalar polarizability of every dipole.

alpha_m: a three-dimensional complex array of size $N\times 3\times 3$containing the magnetic polarisability $3\times 3$ tensor of every dipole, or one dimensional array of size $N$ containing the scalar polarizability of every dipole.

input_field: a function taking the wavenumber and one position vector of length 3, i.e. of the form field(knorm,r), and that output a complex array of length 3 which computes the input fields $E_0$ and $H_0$ evaluated at this position. Can be for example a plane wave or a point source.

solver: a string that five the name of the solver that need to be used. For this, check the correponding section on the home page. By default set to "LAPACK".

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

#Outputs

$6N\times 6N$ inverse of the DDA matrix

Main.CoupledElectricMagneticDipoles.DDACore.solve_DDA_e_m — Method

 solve_DDA_e_m(kr,alpha_e_dl,alpha_m_dl,input_field::Function;solver="AUTO",verbose=true)

Builds and solves the DDA equations with dimensionless inputs under a given input field, i.e.

\[\begin{align} \vec{E}_{i} & =\vec{E}_{0}\left(\vec{r}_{i}\right)+\sum_{j\ne i}\mathbb{G}_{E}\left(\vec{r}_{i},\vec{r}_{j}\right)\alpha_{E}^{(j)}\vec{E}_{j}+i\mathbb{G}_{M}\left(\vec{r}_{i},\vec{r}_{j}\right)\alpha_{M}^{(j)}\vec{H}_{j}\label{eq:100a}\\ \vec{H}_{i} & =\vec{H}_{0}\left(\vec{r}_{i}\right)+\sum_{j\ne i}-i\mathbb{G}_{M}\left(\vec{r}_{i},\vec{r}_{j}\right)\alpha_{E}^{(j)}\vec{E}_{j}+\mathbb{G}_{E}\left(\vec{r}_{i},\vec{r}_{j}\right)\alpha_{M}^{(j)}\vec{H}_{j}\label{eq:100b} \end{align}\]

for a group of $N$ electric and magnetic dipoles and return the polarisations of every particle and incident fields on every particle.

#Arguments

kr: a two-dimensional float array of size $N\times 3$ containing the dimensionless positions $k\vec{r}$ of all the dipoles.

alpha_e_dl: a three-dimensional complex array of size $N\times 3\times 3$containing the dimensionless electric polarisability $3\times 3$ tensor of every dipole, or one dimensional array of size $N$ containing the scalar polarizability of every dipole.

alpha_m_dl: a three-dimensional complex array of size $N\times 3\times 3$containing the dimesnionless magnetic polarisability $3\times 3$ tensor of every dipole, or one dimensional array of size $N$ containing the scalar polarizability of every dipole.

input_field: a function taking the dimensionless position vector of length 3, i.e. of the form field(kr), and that output two complex array of length 3 which computes the input fields $E_0$ and $H_0$ evaluated at this position. Can be for example a plane wave or a point source.

solver: a string that five the name of the solver that need to be used. For this, check the correponding section on the home page. By default set to "LAPACK".

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

#Outputs

p: a two-dimensional complex array of size $N\times 3$ containing the electric polarizations $p$ moments of all the electric dipoles (in units of the electric field).

m: a two-dimensional complex array of size $N\times 3$ containing the magnetic polarizations $m$ moments of all the magnetic dipoles (in units of the electric field).

e_inc: a two-dimensional complex array of size $N\times 3$ containing the incident electric fields $E_inc$ on all the particles (in units of the electric field).

h_inc: a two-dimensional complex array of size $N\times 3$ containing the incident magnetic fields $H_inc$ on all the particles (in units of the electric field).

e_inp: a two-dimensional complex array of size $N\times 3$ containing the input electric fields $E_inc$ on all the particles (in units of the electric field).

h_inp: a two-dimensional complex array of size $N\times 3$ containing the input magnetic fields $H_inc$ on all the particles (in units of the electric field).

Main.CoupledElectricMagneticDipoles.DDACore.solve_DDA_e_m — Method

 solve_DDA_e_m(kr,alpha_e_dl,alpha_m_dl;solver="AUTO",verbose=true)

Similar to solve_DDA_e_m(kr,alpha_e_dl,alpha_m_dl,input_field::Function;solver="AUTO",verbose=true), but without input field. Return then the inverse of the DDA Matrix. #Arguments

kr: a two-dimensional float array of size $N\times 3$ containing the dimensionless positions $k\vec{r}$ of all the dipoles.

alpha_e_dl: a three-dimensional complex array of size $N\times 3\times 3$containing the dimensionless electric polarisability $3\times 3$ tensor of every dipole, or one dimensional array of size $N$ containing the scalar polarizability of every dipole.

alpha_m_dl: a three-dimensional complex array of size $N\times 3\times 3$containing the dimesnionless magnetic polarisability $3\times 3$ tensor of every dipole, or one dimensional array of size $N$ containing the scalar polarizability of every dipole.

solver: a string that five the name of the solver that need to be used. For this, check the correponding section on the home page. By default set to "LAPACK".

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

#Outputs

$6N\times 6N$ inverse of the DDA matrix

Main.CoupledElectricMagneticDipoles.DDACore.solve_system — Method

 solve_system(matrix,vector,solver,verbose)

Solves a system of equations of the type $Ax=b$ wher matrix is $A$ and vector is $b$ using the method solver and returns x. To choose the appropriate solver, please read the informations on the home page.

Main.CoupledElectricMagneticDipoles.DDACore.invert_system — Method

 invert_system(matrix,solver,verbose)

Inverts the square matrix matrix using the solver method To choose the appropriate solver, please read the informations on the home page.

Main.CoupledElectricMagneticDipoles.DDACore.load_dda_matrix_e — Method

 load_dda_matrix_e(kr,alpha_dl,verbose)

Builds the electric only DDA matrix $A=[I-G\alpha]$ with dimensionless postitions kr (two dimensional arrays of size $Nx3$) and dimensionless polarisabilities alpha_dl (a three-dimensional complex array of size $N\times 3\times 3$containing the polarisability $3\times 3$ tensor of each dipole, or one dimensional array of size $N$ containing the scalar polarizability of each dipole.) Returns $3N\times 3N$ complex DDA matrix.

Main.CoupledElectricMagneticDipoles.DDACore.load_dda_matrix_e_m — Method

 load_dda_matrix_e_m(kr,alpha_e_dl,alpha_m_dl,verbose)

Builds the electric and magnetic DDA matrix $A=[I-G\alpha]$ with dimensionless postitions kr (two dimensional arrays of size $N\times 3$) and dimensionless electric and magnetic polarisabilities alpha_e_dl and alpha_m_dl (a three-dimensional complex array of size $N\times 3\times 3$containing the polarisability $3\times 3$ tensor of each dipole, or one dimensional array of size $N$ containing the scalar polarizability of each dipole.) Return $6N\times 6N$ complex DDA matrix