Optical Forces

The Forces module allows to computes optical forces aon electric and magnetic dipoles with deterministic input fields. The list of functions is below. Note that N is the number of dipoles in the system.

Functions Documentation

Main.CoupledElectricMagneticDipoles.Forces.force_e — Method

force_e(knorm,kr,alpha_e_dl, Ainv, e_0, dxe_0, dye_0, dze_0)

Computes the optical forces on a system made out of electric dipoles for deterministic input fields.

#Arguments

knorm: wavenumber in the medium.

kr: 2D float array of size $N\times 3$ containing the dimensionless position $k\mathbf{r}$ of each dipole.

alpha_e_dl: complex dimensionless electric polarizability of each dipole. See the Alphas module documentation for accepted formats.

Ainv: (inverse) DDA matrix.

e_0: 2D complex array of size $N\times 3$ containing the external imput field.

dxe_0: 2D complex array of size $N\times 3$ containing the derivative respect the $k*x$ argument of the external imput field.

dye_0: 2D complex array of size $N\times 3$ containing the derivative respect the $k*y$ argument of the external imput field.

dze_0: 2D complex array of size $N\times 3$ containing the derivative respect the $k*z$ argument of the external imput field.

#Outputs

real(fx): float array of Size $N$ with the value of the force along the $x$-axis at each dipole.

real(fy): float array of Size $N$ with the value of the force along the $y$-axis at each dipole.

real(fz): float array of Size $N$ with the value of the force along the $z$-axis at each dipole.

Main.CoupledElectricMagneticDipoles.Forces.force_e_m — Method

force_e_m(k,kr,alpha_dl, Ainv, e_0, dxe_0, dye_0, dze_0)

Same as force_e_m(knorm,kr,alpha_e_dl, alpha_m_dl, Ainv, e_0, dxe_0, dye_0, dze_0), but the electric and magnetic polarizabilities of each dipole are given by a single 6x6 complex matrix. See the Alphas module documentation for accepted formats.

Main.CoupledElectricMagneticDipoles.Forces.force_e_m — Method

force_e_m(knorm,kr,alpha_e_dl, alpha_m_dl, Ainv, e_0, dxe_0, dye_0, dze_0)

Computes the optical forces on a system made out of electric and magnetic dipoles for deterministic input fields.

Arguments

knorm: wavenumber in the medium.

kr: 2D float array of size $N\times 3$ containing the dimensionless position $k\mathbf{r}$ of each dipole.

alpha_e_dl: complex dimensionless electric polarizability of each dipole. See the Alphas module documentation for accepted formats.

alpha_m_dl: complex dimensionless magnetic polarizability of each dipole. See the Alphas module documentation for accepted formats.

Ainv: (inverse) DDA matrix.

e_0: 2D complex array of size $N\times 6$ containing the external imput field.

dxe_0: 2D complex array of size $N\times 6$ containing the derivative respect the $k*x$ argument of the external imput field.

dye_0: 2D complex array of size $N\times 6$ containing the derivative respect the $k*y$ argument of the external imput field.

dze_0: 2D complex array of size $N\times 6$ containing the derivative respect the $k*z$ argument of the external imput field.

Outputs

real(fx): float array of Size $N$ with the value of the force along the $x$-axis at each dipole.

real(fy): float array of Size $N$ with the value of the force along the $y$-axis at each dipole.

real(fz): float array of Size $N$ with the value of the force along the $z$-axis at each dipole.