Geometries library

This module provides two functions to discretize spheres on a cubic lattice.

Functions Documentation

Main.CoupledElectricMagneticDipoles.Geometries.gen_sphere_lattice_cubes — Function

gen_sphere_lattice_cubes(s,r;verbose=true)

Discretizes a sphere of radius R in cubes on a cubic lattice where the radius of this sphere is divided in s lattice constants. ref_id is the refractive index id (we suppose that the entire sphere has the same refractive index.) Returns 2D array with first dimension is the N units that compose the sphere and the second is

x-component of the position of the unit

y-component of the position of the unit

z-component of the position of the unit

distance from the origin

refractive index id

lattice parameter

volume of the cubic unit

Main.CoupledElectricMagneticDipoles.Geometries.gen_sphere_lattice_spheres — Function

gen_sphere_lattice_spheres(s,r;verbose=true)

Discretizes a sphere of radius R in spheres on a cubic lattice where the radius of this sphere is divided in s lattice constants. ref_id is the refractive index id (we suppose that the entire sphere has the same refractive index.) Returns 2D array with first dimension is the N units that compose the sphere and the second is

x-component of the position of the unit

y-component of the position of the unit

z-component of the position of the unit

distance from the origin

refractive index id

lattice parameter

volume of the spherical unit