Green Tensors

This module is aimed to compute the electric and magnetic green tensors for the two system of units.

Functions Documentation

Main.CoupledElectricMagneticDipoles.GreenTensors.G_e — Function

G_e(r1,r2,knorm)

Compute the electric green tensor between two position r1 and r2 with wavenumber knorm. The output isd a 3x3 complex matrix. The electric green tensor is defined as:

\[\tilde{G}_e\left(\vec{r_1},\vec{r_2},k\right)=\frac{e^{ikr}}{4 \pi r}k\left(\frac{(kr)^2+ikr-1}{(kr)^2}I+\frac{-(kr)^2-3ikr+3}{(kr)^2}\vec{u_r}\vec{u_r}\right)\]

with

\[r=|r_1-r_2|, \vec{u_r}=\left(r_1-r_2\right)/r\]

Main.CoupledElectricMagneticDipoles.GreenTensors.dxG_e — Function

dxG_e(r1,r2,knorm)

Compute the derivative magnetic green tensor regarding the x component of r1 between two position r1 and r2 with wavenumber knorm. The output isd a 3x3 complex matrix

Main.CoupledElectricMagneticDipoles.GreenTensors.dyG_e — Function

dyG_e(r1,r2,knorm)

Compute the derivative magnetic green tensor regarding the y component of r1 between two position r1 and r2 with wavenumber knorm. The output isd a 3x3 complex matrix

Main.CoupledElectricMagneticDipoles.GreenTensors.dzG_e — Function

dzG_e(r1,r2,knorm)

Compute the derivative magnetic green tensor regarding the z component of r1 between two position r1 and r2 with wavenumber knorm. The output isd a 3x3 complex matrix

Main.CoupledElectricMagneticDipoles.GreenTensors.G_m — Function

G_m(r1,r2,knorm)

Compute the magnetic green tensor between two position r1 and r2 with wavenumber knorm. The output isd a 3x3 complex matrix. The magnetic green tensor is defined as:

\[\tilde{G}_m\left(\vec{r_1},\vec{r_2},k\right)=\frac{e^{ikr}}{4 \pi r}k\left(\frac{ikr-1}{ikr}\right)\vec{u_r}\]

with

\[r=|r_1-r_2|, \vec{u_r}=\left(r_1-r_2\right)/r\]

Main.CoupledElectricMagneticDipoles.GreenTensors.dxG_m — Function

dxG_m(r1,r2,knorm)

Compute the derivative magnetic green tensor regarding the x component of r1 between two position r1 and r2 with wavenumber knorm. The output isd a 3x3 complex matrix

Main.CoupledElectricMagneticDipoles.GreenTensors.dyG_m — Function

dyG_m(r1,r2,knorm)

Compute the derivative magnetic green tensor regarding the y component of r1 between two position r1 and r2 with wavenumber knorm. The output isd a 3x3 complex matrix

Main.CoupledElectricMagneticDipoles.GreenTensors.dzG_m — Function

dzG_m(r1,r2,knorm)

Compute the derivative magnetic green tensor regarding the z component of r1 between two position r1 and r2 with wavenumber knorm. The output isd a 3x3 complex matrix

Main.CoupledElectricMagneticDipoles.GreenTensors.G_e_renorm — Function

G_e_renorm(kr1,kr2)

Compute the electric green tensor in renormalized units (see Home page) between two position multiplied by the wave number kr1 and kr2 (->dimensionless input). The output is a 3x3 complex matrix. The renormalized electric green tensor is defined as:

\[G_e=\frac{4*\pi}{k^2}\tilde{G}_e\]

Main.CoupledElectricMagneticDipoles.GreenTensors.G_m_renorm — Function

G_m_renorm(kr1,kr2)

Compute the magnetic green tensor in renormalized units (see Home page) between two position multiplied by the wave number kr1 and kr2 (->dimensionless input). The output isd a 3x3 complex matrix. The renormalized magnetic green tensor is defined as:

\[G_m=\frac{4*\pi}{k}\tilde{G}_m\]

Main.CoupledElectricMagneticDipoles.GreenTensors.G_e_m — Function

G_e_m(r1,r2,knorm)

Compute the big electric and magnetic green tensor between two position r1 and r2 with wavenumber knorm. The output is a 6x6 complex matrix. The big electric and magnetic green tensor is defined as:

\[ \tilde{G}_{em}=\left(\begin{matrix} \tilde{G}_e & \tilde{G}_m\\ \tilde{G}_m & \tilde{G}_e \end{matrix}\right)\]

Main.CoupledElectricMagneticDipoles.GreenTensors.G_e_m_renorm — Function

G_e_m_renorm(r1,r2,knorm)

Compute the big electric and magnetic green tensor in renormalized units (see Home page) between two position r1 and r2 with wavenumber knorm. The output is a 6x6 complex matrix. The big electric and magnetic green tensor in renormalized units is defined as:

\[ G_{em}=\left(\begin{matrix} G_e & G_m\\ G_m & G_e \end{matrix}\right)\]

Main.CoupledElectricMagneticDipoles.GreenTensors.dxG_e_m — Function

dxG_e_m(r1,r2,knorm)

Compute the derivative of the big electric and magnetic green tensor regarding the x component of r1 between two position r1 and r2 with wavenumber knorm. The output isd a 6x6 complex matrix

Main.CoupledElectricMagneticDipoles.GreenTensors.dyG_e_m — Function

dyG_e_m(r1,r2,knorm)

Compute the derivative of the big electric and magnetic green tensor regarding the y component of r1 between two position r1 and r2 with wavenumber knorm. The output isd a 6x6 complex matrix

Main.CoupledElectricMagneticDipoles.GreenTensors.dzG_e_m — Function

dzG_e_m(r1,r2,knorm)

Compute the derivative of the big electric and magnetic green tensor regarding the z component of r1 between two position r1 and r2 with wavenumber knorm. The output isd a 6x6 complex matrix