Green Tensors

This module is aimed to compute the electric and magnetic green tensors for the two system of units.

Functions Documentation

Main.CoupledElectricMagneticDipoles.GreenTensors.G_e — Function

G_e(r1,r2,knorm)

Compute the electric green tensor between two position r1 and r2 with wavenumber knorm. The output is a 3x3 complex matrix. The electric green tensor (with units [1/m]) is defined as:

\[\tilde{G}_e=\left(\vec{r_1},\vec{r_2},k\right)=\frac{e^{ikr}}{4 \pi r}\left(\frac{(kr)^2+ikr-1}{(kr)^2}I+\frac{-(kr)^2-3ikr+3}{(kr)^2}\vec{u_r}\otimes\vec{u_r}\right)\]

with

\[r=|r_1-r_2|, \vec{u_r}=\left(r_1-r_2\right)/r\]

Main.CoupledElectricMagneticDipoles.GreenTensors.G_m — Function

G_m(r1,r2,knorm)

Compute the magnetic green tensor between two position r1 and r2 with wavenumber knorm. The output is a 3x3 complex matrix. The magnetic green tensor (with units [1/m^2]) is defined as:

\[\tilde{G}_m=\left(\vec{r_1},\vec{r_2},k\right)=\frac{e^{ikr}}{4 \pi r}k\left(\frac{ikr-1}{ikr}\right)\vec{u_r}\]

with

\[r=|r_1-r_2|, \vec{u_r}=\left(r_1-r_2\right)/r\]

Main.CoupledElectricMagneticDipoles.GreenTensors.G_e_renorm — Function

G_e_renorm(kr1,kr2)

Compute the electric green tensor in renormalized units (see Home page) between two position multiplied by the wave number kr1 and kr2 (->dimensionless input). The output is a 3x3 complex matrix. The renormalized electric green tensor (with units [1]) is defined as:

\[G_e=\frac{4*\pi}{k}\tilde{G}_e\]

Main.CoupledElectricMagneticDipoles.GreenTensors.G_m_renorm — Function

G_m_renorm(kr1,kr2)

Compute the magnetic green tensor in renormalized units (see Home page) between two position multiplied by the wave number kr1 and kr2 (->dimensionless input). The output isd a 3x3 complex matrix. The renormalized magnetic green tensor (with units [1]) is defined as:

\[G_m=\frac{4*\pi}{k^2}\tilde{G}_m\]

Main.CoupledElectricMagneticDipoles.GreenTensors.G_em — Function

G_em(r1,r2,k)

Compute the electric and magnetic green tensor between two position r1 and r2 (r2 the origin and r1 the observation) . The output are two 3x3 complex matrix. The electric green tensor (with unit of [1/m]) is defined as:

\[\tilde{G}_e=\left(\vec{r_1},\vec{r_2},k\right)=\frac{e^{ikr}}{4 \pi r}\left(\frac{(kr)^2+ikr-1}{(kr)^2}I+\frac{-(kr)^2-3ikr+3}{(kr)^2}\vec{u_r}\otimes\vec{u_r}\right)\]

The magnetic green tensor (with unit of [1/m]) is defined as:

\[\tilde{G}_m/k=\left(\vec{r_1},\vec{r_2},k\right)=\frac{e^{ikr}}{4 \pi r}\left(\frac{ikr-1}{ikr}\right)\vec{u_r}\]

Main.CoupledElectricMagneticDipoles.GreenTensors.G_em_renorm — Function

G_em_renorm(kr1,kr2)

Compute the electric and magnetic green tensor between two position r1 and r2, where the imputs are the positions multiplied by the wave number kr1 and kr2 (->dimensionless input). The output are two 3x3 complex matrix. The electric green tensor (with unit of [1]) is defined as:

\[G_e=\frac{4*\pi}{k}\tilde{G}_e\]

The magnetic green tensor (with unit of [1]) is defined as:

\[G_m=\frac{4*\pi}{k^2}\tilde{G}_m\]

Main.CoupledElectricMagneticDipoles.GreenTensors.dxG_e — Function

dxG_e(r1,r2,knorm)

Compute the derivative electric green tensor (defined in G_e(r1,r2,knorm)) regarding the x component of r1 between two position r1 and r2 with wavenumber knorm. The output isd a 3x3 complex matrix with units [1/m^2]

Main.CoupledElectricMagneticDipoles.GreenTensors.dyG_e — Function

dyG_e(r1,r2,knorm)

Compute the derivative electric green tensor (defined in G_e(r1,r2,knorm)) regarding the y component of r1 between two position r1 and r2 with wavenumber knorm. The output isd a 3x3 complex matrix with units [1/m^2]

Main.CoupledElectricMagneticDipoles.GreenTensors.dzG_e — Function

dzG_e(r1,r2,knorm)

Compute the derivativeelectric green tensor (defined in G_e(r1,r2,knorm)) regarding the z component of r1 between two position r1 and r2 with wavenumber knorm. The output isd a 3x3 complex matrix with units [1/m^2]

Main.CoupledElectricMagneticDipoles.GreenTensors.dxG_m — Function

dxG_m(r1,r2,knorm)

Compute the derivative magnetic green tensor (defined in G_m(r1,r2,knorm)) regarding the x component of r1 between two position r1 and r2 with wavenumber knorm. The output isd a 3x3 complex matrix with units [1/m^3]

Main.CoupledElectricMagneticDipoles.GreenTensors.dyG_m — Function

dyG_m(r1,r2,knorm)

Compute the derivative magnetic green tensor (defined in G_m(r1,r2,knorm)) regarding the y component of r1 between two position r1 and r2 with wavenumber knorm. The output isd a 3x3 complex matrix with units [1/m^3]

Main.CoupledElectricMagneticDipoles.GreenTensors.dzG_m — Function

dzG_m(r1,r2,knorm)

Compute the derivative magnetic green tensor (defined in G_m(r1,r2,knorm)) regarding the z component of r1 between two position r1 and r2 with wavenumber knorm. The output isd a 3x3 complex matrix with units [1/m^3]

Main.CoupledElectricMagneticDipoles.GreenTensors.dxG_em — Function

dxG_em(r1,r2,knorm)

Compute the derivative green tensor (defined in Ge(r1,r2,knorm) and Gm(r1,r2,knorm)/knorm) regarding the x component of r1 between two position r1 and r2 with wavenumber knorm. The output are two 3x3 complex matrix (with units of [1/m^2]).

Main.CoupledElectricMagneticDipoles.GreenTensors.dyG_em — Function

dyG_em(r1,r2,knorm)

Compute the derivative green tensor (defined in Ge(r1,r2,knorm) and Gm(r1,r2,knorm)/knorm) regarding the y component of r1 between two position r1 and r2 with wavenumber knorm. The output are two 3x3 complex matrix (with units of [1/m^2]).

Main.CoupledElectricMagneticDipoles.GreenTensors.dzG_em — Function

dzG_em(r1,r2,knorm)

Compute the derivative green tensor (defined in Ge(r1,r2,knorm) and Gm(r1,r2,knorm)/knorm) regarding the x component of r1 between two position r1 and r2 with wavenumber knorm. The output are two 3x3 complex matrix (with units of [1/m^2]).

Main.CoupledElectricMagneticDipoles.GreenTensors.Sigma — Function

Sigma(n)

Sigma represent the base-change-matrix [E, ZH] to [E, iZH]