Input Fields

This module is aimed to compute the electric and magnetic green tensors for the two system of units.

Functions Documentation

Main.CoupledElectricMagneticDipoles.InputFields.plane_wave_e — Function

plane_wave_e(kr;khat=[0,0,1],e0=[1,0,0])

Computes a simple plane with dimensionless input evaluated at kr. khat is the direction of propagation and e0 is the polarization. kr is a $N\times 3$ float array. The output is a $N\times 3$ complex array represeinting the electric field. This plane wave is defined as:

\[\mathbf{E}\left(\mathbf{r}\right)=\mathbf{E}_{0}e^{i\mathbf{k}\cdot\mathbf{r}}\]

Main.CoupledElectricMagneticDipoles.InputFields.plane_wave_e_m — Function

plane_wave_e_m(kr;khat=[0,0,1],e0=[1,0,0])

Computes a simple plane with dimensionless input evaluated at kr. khat is the direction of propagation and e0 is the polarization. kr is a $N\times 3$ float array. The output is a $N\times 6$ complex array representing the electric and magnetic field. This plane wave is defined as:

\[\mathbf{E}\left(\mathbf{r}\right)=\mathbf{E}_{0}e^{i\mathbf{k}\cdot\mathbf{r}}\]

Main.CoupledElectricMagneticDipoles.InputFields.point_dipole_e — Function

point_dipole_e(krf, krd, dip, e0_const=1)

Function that calculated the electromagnetic field emitted by a point dipole.

#Arguments

krf: 2D float array of size $N\times 3$ containing the dimentionless positions $k\vec{r_f}$ where the field is calculated.

krd: 2D float array of size $1\times 3$ containing the dimentionless positions $k\vec{r_d}$ where source is located.

dip: integer defining the dipole moment ($dip = 1$ is an electric x-dipole, $dip = 2$ an elctric y-dipole...) or float array of size 3 with the desired dipole moment of the dipole.

#Outputs

e_dipole: complex array with the electromagnetic field.

Main.CoupledElectricMagneticDipoles.InputFields.point_dipole_e_m — Function

point_dipole_e_m(krf, krd, dip, e0_const=1)

Function that calculated the electromagnetic field emitted by a point dipole.

#Arguments

krf: 2D float array of size $N\times 3$ containing the dimentionless positions $k\vec{r_f}$ where the field is calculated.

krd: 2D float array of size $1\times 3$ containing the dimentionless positions $k\vec{r_d}$ where source is located.

dip: integer defining the dipole moment ($dip = 1$ is an electric x-dipole, $dip = 2$ an elctric y-dipole...) or float array of size 6 with the desired dipole moment of the dipole.

#Outputs

e_dipole: complex array with the electromagnetic field.

Main.CoupledElectricMagneticDipoles.InputFields.gauss_beam_e — Function

gauss_beam_e(rf,k,bw0, maxe = Int(5e3))

Field distribution of a Gaussian beam that propagates along the z-axis and the electric field is polarized along the x-axis (polarized electric). For another polarization just rotate the field in the .xy-plane. Also, for a polarized magnetic field, exchange E -> Hz and Hz -> -E.

#Arguments

rf: 2D float array of size $N\times 3$ containing the positions where the field is calculated ($N$ is the number of positions).

k: scalar with the modulus wavevector.

bw0: float with the beam waist radius.

maxe: maximum number of evaluations in the adapative integral (see Cubature for more details).

#Outputs

eh_gauss: 2D complex array of size $N\times 6$ with the value of the field at every position.

Main.CoupledElectricMagneticDipoles.InputFields.ghermite_beam_e — Function

ghermite_beam_e(rf,k,bw0,n,m,maxe=Int(5e3))

Field distribution of a Hermite-Gaussian beam that propagates along the z-axis and the electric field is polarized along the x-axis (polarized electric). For another polarization just rotate the field in the .xy-plane. Also, for a polarized magnetic field, exchange E -> Hz and Hz -> -E.

#Arguments

rf: 2D float array of size $N\times 3$ containing the positions where the field is calculated ($N$ is the number of positions).

k: scalar with the modulus wavevector.

bw0: float with the beam waist radius.

n: int with the order of the beam.

m: int with the degree of the beam.

maxe: maximum number of evaluations in the adapative integral (see Cubature for more details).

#Outputs

eh_hermite: 2D complex array of size $N\times 6$ with the value of the field at every position.

Main.CoupledElectricMagneticDipoles.InputFields.glaguerre_beam_e — Function

glaguerre_beam_e(rf,k,bw0,n,m,maxe=Int(5e3))

Field distribution of a Laguerre-Gaussian beam that propagates along the z-axis and the electric field is polarized along the x-axis (polarized electric). For another polarization just rotate the field in the .xy-plane. Also, for a polarized magnetic field, exchange E -> Hz and Hz -> -E.

#Arguments

rf: 2D float array of size $N\times 3$ containing the positions where the field is calculated ($N$ is the number of positions).

k: scalar with the modulus wavevector.

bw0: float with the beam waist radius.

n: non-negative int with the radial order of the beam.

m: int with the azimuthal order of the beam.

maxe: maximum number of evaluations in the adapative integral (see Cubature for more details).

#Outputs

eh_hermite: 2D complex array of size $N\times 6$ with the value of the field at every position.

Main.CoupledElectricMagneticDipoles.InputFields.gauss_beam_e_m — Function

gauss_beam_e_m(rf,k,bw0, maxe = Int(5e3))

Field distribution of a Gaussian beam that propagates along the z-axis and the electric field is polarized along the x-axis (polarized electric). For another polarization just rotate the field in the .xy-plane. Also, for a polarized magnetic field, exchange E -> Hz and Hz -> -E.

#Arguments

rf: 2D float array of size $N\times 3$ containing the positions where the field is calculated ($N$ is the number of positions).

k: scalar with the modulus wavevector.

bw0: float with the beam waist radius.

maxe: maximum number of evaluations in the adapative integral (see Cubature for more details).

#Outputs

eh_gauss: 2D complex array of size $N\times 6$ with the value of the field at every position.

Main.CoupledElectricMagneticDipoles.InputFields.ghermite_beam_e_m — Function

ghermite_beam_e_m(rf,k,bw0,n,m,maxe=Int(5e3))

Field distribution of a Hermite-Gaussian beam that propagates along the z-axis and the electric field is polarized along the x-axis (polarized electric). For another polarization just rotate the field in the .xy-plane. Also, for a polarized magnetic field, exchange E -> Hz and Hz -> -E.

#Arguments

rf: 2D float array of size $N\times 3$ containing the positions where the field is calculated ($N$ is the number of positions).

k: scalar with the modulus wavevector.

bw0: float with the beam waist radius.

n: int with the order of the beam.

m: int with the degree of the beam.

maxe: maximum number of evaluations in the adapative integral (see Cubature for more details).

#Outputs

eh_hermite: 2D complex array of size $N\times 6$ with the value of the field at every position.

Main.CoupledElectricMagneticDipoles.InputFields.glaguerre_beam_e_m — Function

glaguerre_beam_e_m(rf,k,bw0,n,m,maxe=Int(5e3))

Field distribution of a Laguerre-Gaussian beam that propagates along the z-axis and the electric field is polarized along the x-axis (polarized electric). For another polarization just rotate the field in the .xy-plane. Also, for a polarized magnetic field, exchange E -> Hz and Hz -> -E.

#Arguments

rf: 2D float array of size $N\times 3$ containing the positions where the field is calculated ($N$ is the number of positions).

k: scalar with the modulus wavevector.

bw0: float with the beam waist radius.

n: non-negative int with the radial order of the beam.

m: int with the azimuthal order of the beam.

maxe: maximum number of evaluations in the adapative integral (see Cubature for more details).

#Outputs

eh_hermite: 2D complex array of size $N\times 6$ with the value of the field at every position.