Input Fields

This module is aimed to compute the electric and magnetic green tensors for the two system of units.

Module Features

Computes the electric/magnetic green tensor.

Computes the electric/magnetic green tensor for renormalized units.

Compute the big electric and magnetic green tensor.

Compute the big electric and magnetic for renormalized units.

Compute the derivatives of the electric/magnetic/big green tensor.

Function Documentation

Main.input_fields.plane_wave — Function

plane_wave(knorm,r,khat=[0,0,1],e0=[1,0,0])

Computes a simple plane wave of wavenumber knorm evaluated at position r function. khat is the direction of propagation and e0 is the polarization. The output is a 3d complex vector. This plane wave is defined as:

\[\mathbf{E}\left(\mathbf{r}\right)=\mathbf{E}_{0}e^{i\mathbf{k}\cdot\mathbf{r}}\]