Mie Coefficients

MieCoeff is a small module for computing Mie Coefficients. It also contains some useful function to compute cross sections (scattering efficiencies) from the Mie theory.

Expressions are taken in Craig F. Bohren, Donald R. Huffman, Absorption and Scattering of Light by Small Particles (1998).

Functions Documentation

Main.CoupledElectricMagneticDipoles.MieCoeff.mie_an — Method

mie_an(ka, eps, eps_h; mu=1, mu_h=1, n=1)

Computes the n-th mie coefficient $a_n$ of a sphere with size parameter ka, dielectric permittivity eps and magnetic permeability mu, in a host medium with dielectric permittivity eps_h and magnetic permeability mu_h. Returns a complex scalar.

Main.CoupledElectricMagneticDipoles.MieCoeff.mie_bn — Method

mie_bn(ka, eps, eps_h; mu=1, mu_h=1, n=1)

Computes the n-th mie coefficient $b_n$ of a sphere with size parameter ka, dielectric permittivity eps and magnetic permeability mu, in a host medium with dielectric permittivity eps_h and magnetic permeability mu_h. Returns a complex scalar.

Main.CoupledElectricMagneticDipoles.MieCoeff.mie_ab1 — Method

mie_ab1(ka, eps, eps_h; mu=1, mu_h=1)

Computes the first mie coefficient $a_1$ and $b_1$ of a sphere with size parameter ka, dielectric permittivity eps and magnetic permeability mu, in a host medium with dielectric permittivity eps_h and magnetic permeability mu_h. Returns a tuple with two complex scalar, $a_1$ and $b_1$, respectively.

Main.CoupledElectricMagneticDipoles.MieCoeff.mie_scattering — Method

mie_scattering(ka,eps,eps_h;mu=1, mu_h=1, cutoff=20)

Computes the scattering efficiency $Q_{sca}$ of a sphere with size parameter ka, dielectric permittivity eps and magnetic permeability mu, in a host medium with dielectric permittivity eps_h and magnetic permeability mu_h. For this, we use the finite sum:

\[Q_{sca} =\frac{2}{ka^2}\sum^{\text{cutoff}}_{n=1}\left(2n+1\right)\left(|a_n|^2+|b_n|^2\right)\]

where cutoff is set to 20 by default.

Returns a float.

Main.CoupledElectricMagneticDipoles.MieCoeff.mie_extinction — Method

mie_extinction(ka,eps,eps_h;cutoff=20)

Computes the extinction efficiency $Q_{ext}$ of a sphere with size parameter ka, dielectric permittivity eps and magnetic permeability mu, in a host medium with dielectric permittivity eps_h and magnetic permeability mu_h. For this, we use the finite sum:

\[Q_{ext} =\frac{2}{(ka)^2}\sum^{\text{cutoff}}_{n=1}\left(2n+1\right)Re\left(a_n+b_n\right)\]

where cutoff is set to 20 by default.

Returns a float.

Main.CoupledElectricMagneticDipoles.MieCoeff.mie_absorption — Method

mie_absorption(ka,eps,eps_h;mu=1,mu_h=1,cutoff=20)

Computes the extinction efficiency $Q_{abs}$ of a sphere with size parameter ka, dielectric permittivity eps and magnetic permeability mu, in a host medium with dielectric permittivity eps_h and magnetic permeability mu_h. For this, we use:

\[Q_{abs} =Q_{ext}-Q_{sca}\]

where corresponding sums are cut to cutoff which is set to 20 by default.

Returns a float.