Post Processing

Functions Documentation

Missing docstring.

Missing docstring for CoupledElectricMagneticDipoles.PostProcessing.compute_cross_sections. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CoupledElectricMagneticDipoles.PostProcessing.compute_cross_sections_e_m. Check Documenter's build log for details.

Main.CoupledElectricMagneticDipoles.PostProcessing.point_dipole — Function

point_dipole(knorm, E0_const, positions, rd, dip_o)

Function that calculated the electromagnetic field emitted by a point dipole with dipole moment

\[\mathrm{dip}_o = \dfrac{1}{\epsilon_0\epsilon} \overrightarrow{\mu}, \quad \mathrm{(see \ equation \ below)}\]

Imputs

knorm0 is the medium wavevector (scalar)

E0_const is the field intensity (scalar). The modulus of the dipole moment is set to -epsilon0*epsilonm-, where -epsilon0- and -epsilonm- are the vacuum and medium permittivity, respectively.

position contains the position at which the field is calculated (N x 3 matrix, where -N- is the number of points)

rd is the position of the emitting source/dipole (1 x 3 vector)

dip_o defined the nature of the dipole. If dip_o is a scalar then:

dip_o = 1 -> elecric dipole along x-axis

dip_o = 2 -> elecric dipole along y-axis

dip_o = 3 -> elecric dipole along z-axis

dip_o = 4 -> magnetic dipole along x-axis

dip_o = 5 -> magnetic dipole along y-axis

dip_o = 6 -> magnetic dipole along z-axis

if dip_o is a 6 x 1 vector then it specifies the dipole moment orentation of the source.

Outputs

E_0i is the electromagnetic field vector of the field at the requiered positions (6N x 1 vector)

Equation

\[\mathbf{E}_{\mathbf{\mu}}(\mathbf{r}) = \omega^2 \mu \mu_0 G(\mathbf{r}, \mathbf{r}_0) \overrightarrow{\mu} = k^2 G(\mathbf{r}, \mathbf{r_0}) \dfrac{1}{\epsilon_0\epsilon} \overrightarrow{\mu}\]

with

\[\mathrm{positions} = \mathbf{r}, \\ \mathrm{rd} = \mathbf{r}_0, \\ \mathrm{E}_{\mathrm{0i}} = \mathbf{E}_{\mathbf{\mu}}(\mathbf{r}).\]

Main.CoupledElectricMagneticDipoles.PostProcessing.point_dipole_dl — Function

point_dipole_dl(knorm, E0_const, kpositions, krd, dip_o)

Function that calculated the electromagnetic field emitted by a point dipole with dipole moment

\[\mathrm{dip}_o = \dfrac{1}{\epsilon_0\epsilon} \overrightarrow{\mu}, \quad \mathrm{(see \ equation \ below)}\]

Imputs

knorm0 is the medium wavevector (scalar)

E0_const is the field intensity (scalar). The modulus of the dipole moment is set to -epsilon0*epsilonm-, where -epsilon0- and -epsilonm- are the vacuum and medium permittivity, respectively.

kposition contains the position (multiplied by the wavevector) at which the field is calculated (N x 3 matrix, where -N- is the number of points)

krd is the position of the emitting source/dipole (multiplied by the wavevector) (1 x 3 vector)

dip_o defined the nature of the dipole. If dip_o is a scalar then:

dip_o = 1 -> elecric dipole along x-axis

dip_o = 2 -> elecric dipole along y-axis

dip_o = 3 -> elecric dipole along z-axis

dip_o = 4 -> magnetic dipole along x-axis

dip_o = 5 -> magnetic dipole along y-axis

dip_o = 6 -> magnetic dipole along z-axis

if dip_o is a 6 x 1 vector then it specifies the dipole moment orentation of the source.

Outputs

E_0i is the electromagnetic field vector of the field at the requiered positions (6N x 1 vector)

Equation

\[\mathbf{E}_{\mathbf{\mu}}(\mathbf{r}) = k^2 G(k\mathbf{r}, k\mathbf{r_0}) \dfrac{1}{\epsilon_0\epsilon} \overrightarrow{\mu}\]

with

\[\mathrm{kpositions} = \mathbf{kr}, \\ \mathrm{krd} = \mathbf{kr}_0, \\ \mathrm{E}_{\mathrm{0i}} = \mathbf{E}_{\mathbf{\mu}}(\mathbf{r}).\]

Main.CoupledElectricMagneticDipoles.PostProcessing.field_sca — Function

field_sca(knorm, alpha, E_inc, r0, pos)

It computes the scattered Field from the ensamble of dipoles.

Imputs

knorm = wavenumber

alpha = polarizability of the particles (6N x 6N matrix, where -N- is the number of dipoles)

E_inc = incoming field at every dipole (6N x 1 vector, where -N- is the number of dipoles). It is equal to the product of the (inverse) DDA matrix and the external field.

r0 = position where the field is observed (Np x 3 matrix, where -Np- is the number of points where the field is calculated)

pos = position of the dipoles (N x 3 matrix, where -N- is the number of points)

Outputs

field_r is the field scattered by the dipoles (6N x 1 vector)

Equation

\[\mathbf{E}_{sca}(\mathbf{r}) = k^2G(\mathbf{r},\mathbf{\bar{r}}_N) \alpha(\mathbf{\bar{r}}_N) \mathbf{E}_{inc}(\mathbf{\bar{r}}_N) = k^2 G(\mathbf{r},\mathbf{\bar{r}}_N) \alpha(\mathbf{\bar{r}}_N) D(\mathbf{\bar{r}}_N) \E_{0}\]

\r = r0 \bar{r}}N = pos \alphagg(\mathbf{\bar{r}}N) = alpha \GG(\r,\mathbf{\bar{r}}N) = `Gtensor\E_{inc}(\mathbf{\bar{r}}_N) =E_inc`

\E{sca}(\r) = `fieldr`

Main.CoupledElectricMagneticDipoles.PostProcessing.LDOS_rf — Function

LDOS_rf(knorm, alpha, Ainv, pos, rd, dip_o)

It Computes partial local density of states (LDOS) by the imaginary part of the returning field (rf)

Imputs

knorm = wavenumber

alpha = polarizability of the particles (6N x 6N matrix, where -N- is the number of dipoles)

Ainv = (inverse) DDA matrix (6N x 6N matrix, [I - k^2Galpha]^(-1))

pos = position of the dipoles (N x 3 matrix, where -N- is the number of points)

rd = emitting dipole position, i. e., position where the LDOS is calculated

dip_o defined the nature of the dipole (see -point_dipole- function). Therefore, it defines the component of the LDOS that is calculated

Outputs

LDOS is a scalar with the value of the partial LDOS

Equation

\[\mathrm{LDOS}(\mathbf{\bar{r}}_N,\r_0) = 1 + \dfrac{1}{|\bm \mu|^2 }\dfrac{6\pi}{k^3} \Im\left[\bm \mu^{*} \cdot k^2\GG(\r_0,\mathbf{\bar{r}}_N) \alphagg(\mathbf{\bar{r}}_N) \DD(\mathbf{\bar{r}}_N) k^2 \GG(\mathbf{\bar{r}}_N, \r_0) \bm \mu \right]\]

\alphagg(\mathbf{\bar{r}}N) = alpha \DD(\mathbf{\bar{r}}N) = Ainv = [I - k^2Galpha]^(-1) \bar{r}}N = pos \r0 = rd \dfrac{1}{\epsilon0\epsilon} \bm \mu = `dipo` (the pre-factor -\dfrac{1}{\epsilon_0\epsilon}- dessapears after normalization)

k^2\GG(\r0,\mathbf{\bar{r}}N) \alphagg(\mathbf{\bar{r}}N) \DD(\mathbf{\bar{r}}N) k^2 \GG(\mathbf{\bar{r}}N, \r0) \bm \mu = field_r \DD(\mathbf{\bar{r}}N) k^2 \GG(\mathbf{\bar{r}}N, \r0) \bm \mu = `Einc`

\mathrm{LDOS}(\mathbf{\bar{r}}N,\r0) = LDOS

Main.CoupledElectricMagneticDipoles.PostProcessing.LDOS_sc — Function

LDOS_rf(knorm, alpha, Ainv, pos, rd, dip_o)

It Computes partial local density of states (LDOS) by the scattering cross section (sc)

Imputs

knorm = wavenumber

alpha = polarizability of the particles (6N x 6N matrix, where -N- is the number of dipoles)

Ainv = (inverse) DDA matrix (6N x 6N matrix, [I - k^2Galpha]^(-1))

pos = position of the dipoles (N x 3 matrix, where -N- is the number of points)

rd = emitting dipole position, i. e., position where the LDOS is calculated

dip_o defined the nature of the dipole (see -point_dipole- function). Therefore, it defines the component of the LDOS that is calculated

Outputs

LDOS is a scalar with the value of the partial LDOS

Equation

\[LDOS = \sigma^{sca}/\sigma^{sca}_{0}\]

Main.CoupledElectricMagneticDipoles.PostProcessing.LDOS_EP — Function

LDOS_EP(knorm, alpha, Ainv, pos, rd, dip_o)

It Computes partial local density of states (LDOS) by calculating the emiting power (EP). Due to the integration, the method is not very accurate.

Imputs

knorm = wavenumber

alpha = polarizability of the particles (6N x 6N matrix, where -N- is the number of dipoles)

Ainv = (inverse) DDA matrix (6N x 6N matrix, [I - k^2Galpha]^(-1))

pos = position of the dipoles (N x 3 matrix, where -N- is the number of points)

rd = emitting dipole position, i. e., position where the LDOS is calculated

dip_o defined the nature of the dipole (see -point_dipole- function). Therefore, it defines the component of the LDOS that is calculated

Outputs

LDOS is a scalar with the value of the partial LDOS

Equation

\[LDOS = P/P_{0}\]