Post Processing

The PostProcessing modules allows you to process the results of a CEMD calculation. Mostly for:

Computing cross sections and differential emitted power.

Computing scattered fields.

Computing local density of states (LDOS).

The list of functions and a description of how to use them is given in the following section. In general $N$ represents the number of dipoles.

Functions List and Documentation

Functions for Cross Sections and Differential Emitted Power

Main.CoupledElectricMagneticDipoles.PostProcessing.compute_cross_sections_e — Method

compute_cross_sections_e(knorm,kr,e_inc,alpha_e_dl,input_field;explicit_scattering=true,verbose=true)

Computes the extinction, absorbtion and scattering cross section $\sigma_{ext}$, $\sigma_{abs}$, $\sigma_{sca}$ of a system made out of electric dipoles, in a plane wave input field. Note that it should follow the optical theorem, i.e.

\[\sigma_{ext}=\sigma_{abs}+\sigma_{sca}\]

Arguments

knorm: wavenumber in the medium.

kr: 2D float array of size $N\times 3$ containing the dimensionless position $k\mathbf{r}$ of each dipole.

e_inc: 2D complex array of size $N\times 3$ containing the incident electric field $\mathbf{E}_{i}$ on each dipole.

alpha_e_dl: complex dimensionless electric polarizability of each dipole. See the Alphas module documentation for accepted formats.

input_field: 2D complex array of size $N\times 3$ containing the input field $\mathbf{E}_0(\mathbf{r}_i)$ at the position of each dipole. Note that it must be a plane wave.

explicit_scattering: boolean that says whether to compute scttering cross section explicitely (true) or to deduce it from the optical theorem (false). By default set to true.

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

Outputs

a float array of length 3 containing in order: extinction, absorption and scattering cross section.

Main.CoupledElectricMagneticDipoles.PostProcessing.compute_cross_sections_e_m — Method

compute_cross_sections_e_m(knorm,kr,phi_inc,alpha_e_dl,alpha_m_dl,input_field;explicit_scattering=true,verbose=true)

Computes the extinction, absorbtion and scattering cross section $\sigma_{ext}$, $\sigma_{abs}$, $\sigma_{sca}$ of a system made out of electric and magnetic dipoles, in a plane wave input field. Note that it should follow the optical theorem, i.e.

\[\sigma_{ext}=\sigma_{abs}+\sigma_{sca}\]

Arguments

knorm: wavenumber in the medium.

kr: 2D float array of size $N\times 3$ containing the dimensionless position $k\mathbf{r}$ of each dipole.

phi_inc: 2D complex array of size $N\times 6$ containing the incident electric and magnetic field $\mathbf{\phi}=(\mathbf{E}_i,\mathbf{H}_i)$ on each dipole.

alpha_e_dl: complex dimensionless electric polarizability of each dipole. See the Alphas module documentation for accepted formats.

alpha_m_dl: complex dimensionless magnetic polarizability of each dipole. See the Alphas module documentation for accepted formats.

input_field: 2D complex array of size $N\times 6$ containing the electric and magnetic input field $\mathbf{\phi}=(\mathbf{E}_0(\mathbf{r}_i),\mathbf{H}_0(\mathbf{r}_i))$ at the position of each dipole. Note that it should be a plane wave.

explicit_scattering: boolean that says whether to compute scttering cross section explicitely (true) or to deduce it from the optical theorem (false). By default set to true.

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

Outputs

a float array of length 3 containing in order: extinction, absorption and scattering cross section.

Main.CoupledElectricMagneticDipoles.PostProcessing.compute_cross_sections_e_m — Method

compute_cross_sections_e_m(knorm,kr,phi_inc,alpha_dl,input_field;explicit_scattering=true,verbose=true)

Same as compute_cross_sections_e_m(knorm,kr,phi_inc,alpha_e_dl,alpha_m_dl,input_field;explicit_scattering=true,verbose=true), but the electric and magnetic polarizabilities of each dipole are given by a single 6x6 complex matrix. See the Alphas module documentation for accepted formats.

Main.CoupledElectricMagneticDipoles.PostProcessing.diff_scattering_cross_section_e — Method

diff_scattering_cross_section_e(knorm,kr,e_inc,alpha_e_dl,input_field,ur;verbose=true)

Computes the differential scattering cross section $d \sigma_{sca}/ d\Omega$ of a system made out of electric dipoles in direction(s) ur.

Arguments

knorm: wavenumber in the medium.

kr: 2D float array of size $N\times 3$ containing the dimensionless position $k\mathbf{r}$ of each dipole.

e_inc: 2D complex array of size $N\times 3$ containing the incident electric field $\mathbf{E}_{i}$ on each dipole.

alpha_e_dl: complex dimensionless electric polarizability of each dipole. See the Alphas module documentation for accepted formats.

input_field: 2D complex array of size $N\times 3$ containing the electric input field $\mathbf{E}_0(\mathbf{r}_i)$ at the position of each dipole. Note that it should be a plane wave

ur: 1D float vector of length 3 (only one direction) or 2D float array of size $Nu\times 3$ (more thant 1 directions) containing the dimensionless positions $k\mathbf{u_r}$ where the diffrential scattering cross section is computed.

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

Outputs

an array containing the differential cross section in each direction.

Main.CoupledElectricMagneticDipoles.PostProcessing.diff_scattering_cross_section_e_m — Method

diff_scattering_cross_section_e_m(knorm,kr,phi_inc,alpha_e_dl,alpha_m_dl,input_field,ur;verbose=true)

Computes the differential scattering cross section $d \sigma_{sca}/ d\Omega$ of a system made out of electric and magnetic dipoles in direction(s) ur.

Arguments

knorm: wavenumber in the medium.

kr: 2D float array of size $N\times 3$ containing the dimensionless position $k\mathbf{r}$ of each dipole.

phi_inc: 2D complex array of size $N\times 6$ containing the incident electric and magnetic field $\mathbf{\phi}=(\mathbf{E}_i,\mathbf{H}_i)$ on each dipole.

alpha_e_dl: complex dimensionless electric polarizability of each dipole. See the Alphas module documentation for accepted formats.

alpha_m_dl: complex dimensionless magnetic polarizability of each dipole. See the Alphas module documentation for accepted formats.

input_field: 2D complex array of size $N\times 6$ containing the electric and magnetic input field $\mathbf{\phi}=(\mathbf{E}_0(\mathbf{r}_i),\mathbf{H}_0(\mathbf{r}_i))$ at the position of each dipole. Note that it should be a plane wave.

ur: 1D float vector of length 3 (only one direction) or 2D float array of size $Nu\times 3$ (more thant 1 directions) containing the dimensionless positions $k\mathbf{u_r}$ where the diffrential scattering cross section is computed.

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

Outputs

an array containing the differential cross section in each direction.

Main.CoupledElectricMagneticDipoles.PostProcessing.diff_scattering_cross_section_e_m — Method

diff_scattering_cross_section_e_m(knorm,kr,phi_inc,alpha_dl,input_field,ur;verbose=true)

Same as compute_cross_sections_e_m(knorm,kr,phi_inc,alpha_e_dl,alpha_m_dl,input_field;explicit_scattering=true,verbose=true), but the electric and magnetic polarizabilities of each dipole are given by a single 6x6 complex matrix. See the Alphas module documentation for accepted formats.

Main.CoupledElectricMagneticDipoles.PostProcessing.diff_emitted_power_e — Method

diff_emitted_power_e(knorm,kr,e_inc,alpha_e_dl,krf,phi_inp_krf;verbose=true)

Computes the differential emitted power $d \P/ d\Omega$ of a system made out of electric dipoles in direction(s) of position(s) krf.

Arguments

knorm: wavenumber in the medium.

kr: 2D float array of size $N\times 3$ containing the dimensionless position $k\mathbf{r}$ of each dipole.

e_inc: 2D complex array of size $N\times 3$ containing the incident electric field $\mathbf{E}_{i}$ on each dipole.

alpha_e_dl: complex dimensionless electric polarizability of each dipole. See the Alphas module documentation for accepted formats.

input_field: 2D complex array of size $N\times 6$ containing the electric and magnetic input field $\mathbf{\phi}=(\mathbf{E}_0(\mathbf{r}_i),\mathbf{H}_0(\mathbf{r}_i))$ at the position of each dipole.

krf: 1D float vector of length 3 (only one direction) or 2D float array of size $Nu\times 3$ (more thant 1 directions) containing the dimensionless positions $k\mathbf{u_r}$ in which direction the diffrential emitted power is computed. Note that the magnitude of these position must be way bigger than the dipoles positions.

phi_inp_krf: like input_field, but evaluated at the krf positions.

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

Outputs

an array containing the differential emitted power in direction directions.

Main.CoupledElectricMagneticDipoles.PostProcessing.diff_emitted_power_e_m — Method

diff_emitted_power_e_m(knorm,kr,phi_inc,alpha_e_dl,alpha_m_dl,krf,phi_inp_krf;verbose=true)

Computes the differential emitted power $d \P/ d\Omega$ of a system made out of electric and magnetic dipoles in direction(s) of position(s) krf.

Arguments

knorm: wavenumber in the medium.

kr: 2D float array of size $N\times 3$ containing the dimensionless position $k\mathbf{r}$ of each dipole.

phi_inc: 2D complex array of size $N\times 6$ containing the incident electric and magnetic field $\mathbf{\phi}=(\mathbf{E}_i,\mathbf{H}_i)$ on each dipole.

alpha_e_dl: complex dimensionless electric polarizability of each dipole. See the Alphas module documentation for accepted formats.

alpha_m_dl: complex dimensionless magnetic polarizability of each dipole. See the Alphas module documentation for accepted formats.

input_field: 2D complex array of size $N\times 6$ containing the electric and magnetic input field $\mathbf{\phi}=(\mathbf{E}_0(\mathbf{r}_i),\mathbf{H}_0(\mathbf{r}_i))$ at the position of each dipole.

krf: 1D float vector of length 3 (only one direction) or 2D float array of size $Nu\times 3$ (more thant 1 directions) containing the dimensionless positions $k\mathbf{u_r}$ in which direction the diffrential emitted power is computed. Note that the magnitude of these position must be way bigger than the dipoles positions.

phi_inp_krf: like input_field, but evaluated at the krf positions.

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

Outputs

an array containing the differential emitted power in direction directions.

Main.CoupledElectricMagneticDipoles.PostProcessing.diff_emitted_power_e_m — Method

diff_emitted_power_e_m(knorm,kr,phi_inc,alpha_e_dl,alpha_m_dl,krf,phi_inp_krf;verbose=true)

Same as diff_emitted_power_e_m(knorm,kr,phi_inc,alpha_e_dl,alpha_m_dl,krf,phi_inp_krf;verbose=true), but the electric and magnetic polarizabilities of each dipole are given by a single 6x6 complex matrix. See the Alphas module documentation for accepted formats.

Functions for Scattered Fields

Main.CoupledElectricMagneticDipoles.PostProcessing.field_sca_e — Method

function field_sca_e(kr, alpha_e_dl, e_inc, krf)

Computes the scattered field from a system made out of electric dipoles.

Arguments

kr: 2D float array of size $N\times 3$ containing the dimensionless position $k\mathbf{r}$ of each dipole.

alpha_e_dl: complex dimensionless electric polarizability of each dipole. See the Alphas module documentation for accepted formats.

e_inc: 2D complex array of size $N\times 3$ containing the incident electric field $\mathbf{E}_{i}$ on each dipole.

krf: 1D float vector of length 3 (only one position) or 2D float array of size $Nu\times 3$ (more thant 1 position) containing the dimensionless positions $k\mathbf{r}_f$ at which the scattered field is computed.

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

Outputs

field_r: 2D complex array of size $Nf\times 6$ containing the scattered field by the dipoles at every $k\mathbf{r_f}$.

Main.CoupledElectricMagneticDipoles.PostProcessing.field_sca_e_m — Method

field_sca_e_m(kr, alpha_e_dl, alpha_m_dl, phi_inc, krf; verbose=true)

Computes the scattered field from a system made out of electric and magnetic dipoles.

Arguments

kr: 2D float array of size $N\times 3$ containing the dimensionless position $k\mathbf{r}$ of each dipole.

alpha_e_dl: complex dimensionless electric polarizability of each dipole. See the Alphas module documentation for accepted formats.

alpha_m_dl: complex dimensionless magnetic polarizability of each dipole. See the Alphas module documentation for accepted formats.

phi_inc: 2D complex array of size $N\times 6$ containing the incident electric and magnetic field $\mathbf{\phi}=(\mathbf{E}_i,\mathbf{H}_i)$ on each dipole.

krf: 1D float vector of length 3 (only one position) or 2D float array of size $Nu\times 3$ (more thant 1 position) containing the dimensionless positions $k\mathbf{r}_f$ at which the scattered field is computed.

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

Outputs

field_r: 2D complex array of size $Nf\times 6$ containing the scattered field by the dipoles at every $k\mathbf{r_f}$.

Main.CoupledElectricMagneticDipoles.PostProcessing.field_sca_e_m — Method

field_sca_e_m(kr, alpha_dl, phi_inc, krf)

Same as field_sca_e_m(kr, alpha_e_dl, alpha_m_dl, phi_inc, krf), but the electric and magnetic polarizabilities of each dipole are given by a single 6x6 complex matrix. See the Alphas module documentation for accepted formats.

Main.CoupledElectricMagneticDipoles.PostProcessing.far_field_sca_e — Method

function far_field_sca_e(kr,e_inc,alpha_e_dl,krf)

Computes the scattered field from a system made out of electric dipoles in the far field approximation.

Arguments

kr: 2D float array of size $N\times 3$ containing the dimensionless position $k\mathbf{r}$ of each dipole. Note that these positions have to be far away from the dipoles positions.

alpha_e_dl: complex dimensionless electric polarizability of each dipole. See the Alphas module documentation for accepted formats.

e_inc: 2D complex array of size $N\times 3$ containing the incident electric field $\mathbf{E}_{i}$ on each dipole.

krf: 1D float vector of length 3 (only one position) or 2D float array of size $Nu\times 3$ (more thant 1 position) containing the dimensionless positions $k\mathbf{r}_f$ at which the scattered field is computed.

Outputs

field_r: 2D complex array of size $Nf\times 6$ containing the scattered field by the dipoles at every $k\mathbf{r_f}$.

Main.CoupledElectricMagneticDipoles.PostProcessing.far_field_sca_e_m — Method

function far_field_sca_e_m(kr,e_inc,alpha_e_dl,alpha_m_dl,krf)

Computes the scattered field from a system made out of electric and magnetic dipoles in the far field approximation.

Arguments

kr: 2D float array of size $N\times 3$ containing the dimensionless position $k\mathbf{r}$ of each dipole. Note that these positions have to be far away from the dipoles positions.

alpha_e_dl: complex dimensionless electric polarizability of each dipole. See the Alphas module documentation for accepted formats.

alpha_m_dl: complex dimensionless magnetic polarizability of each dipole. See the Alphas module documentation for accepted formats.

phi_inc: 2D complex array of size $N\times 6$ containing the incident electric and magnetic field $\mathbf{\phi}=(\mathbf{E}_i,\mathbf{H}_i)$ on each dipole.

krf: 1D float vector of length 3 (only one position) or 2D float array of size $Nu\times 3$ (more thant 1 position) containing the dimensionless positions $k\mathbf{r}_f$ at which the scattered field is computed.

Outputs

field_r: 2D complex array of size $Nf\times 6$ containing the scattered field by the dipoles at every $k\mathbf{r_f}$.

Main.CoupledElectricMagneticDipoles.PostProcessing.far_field_sca_e_m — Method

function far_field_sca_e_m(kr,e_inc,alpha_dl,krf)

Same as far_field_sca_e_m(kr, alpha_e_dl, alpha_m_dl, phi_inc, krf), but the electric and magnetic polarizabilities of each dipole are given by a single 6x6 complex matrix. See the Alphas module documentation for accepted formats.

Functions for LDOS

Main.CoupledElectricMagneticDipoles.PostProcessing.ldos_e — Method

ldos_e(kr, alpha_e_dl, Ainv, krd; dip=nothing, verbose=true)

It Computes local density of states (LDOS) of a system made out of electric dipoles by the imaginary part of the returning field.

Arguments

kr: 2D float array of size $N\times 3$ containing the dimensionless position $k\mathbf{r}$ of each dipole.

alpha_e_dl: complex dimensionless electric polarizability of each dipole. See the Alphas module documentation for accepted formats.

Ainv: (inverse) DDA matrix.

krd: 2D float array of size $Nd\times 3$ containing the dimentionless positions $k\mathbf{r_d}$ where the LDOS is calculated.

dip: integer defining the dipole moment ($dip = 1$ is an electric x-dipole, $dip = 2$ an elctric y-dipole...) or float array of size 3 with the desired dipole moment of the dipole.

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

Outputs

LDOS: float array with the LDOS.

Main.CoupledElectricMagneticDipoles.PostProcessing.ldos_e_m — Method

ldos_e_m(kr, alpha_e_dl, alpha_m_dl, Ainv, krd; dip=nothing, verbose=true)

Computes local density of states (LDOS) of a system made out of electric and magnetic dipoles by taking the imaginary part of the returning field.

Arguments

kr: 2D float array of size $N\times 3$ containing the dimensionless position $k\mathbf{r}$ of each dipole.

alpha_e_dl: complex dimensionless electric polarizability of each dipole. See the Alphas module documentation for accepted formats.

alpha_m_dl: complex dimensionless magnetic polarizability of each dipole. See the Alphas module documentation for accepted formats.

Ainv: (inverse) DDA matrix.

krd: 2D float array of size $Nd\times 3$ containing the dimentionless positions $k\mathbf{r_d}$ where the LDOS is calculated.

dip: integer defining the dipole moment ($dip = 1$ is an electric x-dipole, $dip = 2$ an elctric y-dipole...) or float array of size 6 with the desired dipole moment of the dipole.

verbose: whether to output pieces of information to the standard output during running or not. By default set to true.

Outputs

LDOS: float array with the LDOS.

Main.CoupledElectricMagneticDipoles.PostProcessing.ldos_e_m — Method

ldos_e_m(kr, alpha_dl, Ainv, krd; dip=nothing, verbose=true)

Same as ldos_e_m(kr, alpha_e_dl, alpha_m_dl, Ainv, krd; dip=nothing, verbose=true), but the electric and magnetic polarizabilities of each dipole are given by a single 6x6 complex matrix. See the Alphas module documentation for accepted formats.

Utilities Functions

Main.CoupledElectricMagneticDipoles.PostProcessing.compute_dipole_moment — Method

compute_dipole_moment(alpha,phi_inc)

Computes the dipole moment (magnetic or electric) of a dipole with polarizability alpha under an incident field phi_inc. alpha can be:

a complex scalar

a 1D complex array of size $N$

a $3\times 3$ or $6\times 6$ complex matrix.

a 3D complex array of size $N\times 3\times 3$ or $N\times 6\times 6$

Main.CoupledElectricMagneticDipoles.PostProcessing.poynting_vector — Method

poynting_vector(phi)

Computes the poynting vector of an electromagnetic field. Input is an electric and magnetic field phi(1D complex Array of length 6). Outputs a 1D float array of length 3.